--- client/src/components/pages/SvgFries 2.js
@@ -1,235 +1,0 @@
-import * as React from "react";
-
-const SvgFries = (props) => (
-  <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 232 466" {...props}>
-    <defs>
-      <style>
-        {
-          ".d{fill:#161616}.i{fill:#4d4d4d}.j{fill:#1a1a1a}.k{fill:#212121}.n{fill:#333}.o{fill:#150b0a}.p{fill:#230f0c}.w{fill:#f5f5f5}.x{fill:#eaeaea}"
-        }
-      </style>
-    </defs>
-    <path
-      d="M81.8 438h46.64S119.11 183.31 162 94.69L121.91 63.9S61.27 277.54 81.8 438Z"
-      style={{
-        fill: "#edc367",
-      }}
-    />
-    <path
-      style={{
-        fill: "#c08b40",
-      }}
-      d="m121.91 63.9 38.25-19.59 41.05 30.78-39.18 19.6-40.12-30.79z"
-    />
-    <path
-      d="M201.21 75.09s-55 48.52-52.24 326.53L128.44 438S118.18 160.92 162 94.69Z"
-      style={{
-        fill: "#d0a851",
-      }}
-    />
-    <path
-      className="d"
-      d="M96 133.3c-2.64 1-4.85-.18-7.18-1.47-2.79-1.53-5.76-2.64-8.66-3.94 7.1-2.12 14.17-4.36 21.31-6.31 4.1-1.12 6.34 2.87 9.14 5s-3.78.25-2.66 2.67c-1-.12-1.62.14-1.87 0-2.45-1.34.94-.93-3.37-.94-2.66-.06-3.97 4.07-6.71 4.99Z"
-    />
-    <path
-      d="M96 133.3c-.35.48-1.17 1.39-.65 1.55 6.5 2 12 7.5 19.08 7.14 4-.2 7.88-1.38 11.83-2 14.89-2.43 29.13 1.51 43.17 6.35a7.53 7.53 0 0 0 7.27-.71c6.79-4.21 13.7-8.2 20.6-12.23 1.38-.8 2.9-3 4.42-1.21 1.79 2.1 4.33 4.08 3.82 7.62a11.53 11.53 0 0 1-1.59 5.23c-2-1.08-1.77-3.88-3.86-5.63-3.36 8.84-11.31 10.48-18.28 13.17a20.27 20.27 0 0 0-13 14c-1.75.23-3.49 2.74-5.24-.05-.25-2.26.64-4.3 1-6.46.73-5-.61-8.66-4.88-10.24-9.71-3.59-19.67-6.13-30-4.56-7.22 1.1-7.77 2-8.09 10-.69 2-.76 4.67-3.79 4.07-.61-2.05-1.24-4.1-1.84-6.16-.88-3-3-3.63-5.46-3.4-2.64.25-1.74 2.62-1.84 4.28-.27 4.38-3 7.48-4.67 11.13-1.45.38-2.89 2.72-4.36 0 .2-3.93 2.17-7.31 2.79-11.13 1-6.44-.05-9.41-5.43-12.42-5.2-2.92-10.7-4.61-16.53-3-4.43 1.22-9.28.53-13.31 3.44-1.37 1.29-.79 5.2-4.19 3.78-.53-1.39-1.16-2.73-1.58-4.16-2.07-7.13-1.54-8.16 5-10 4.6-1.28 9.21-2.5 13.82-3.75 5.7-.07 10.33 4.11 15.79 5.35Z"
-      style={{
-        fill: "#1c1c1c",
-      }}
-    />
-    <path
-      d="M164.12 167.36c1.51-.6 2.91-1.81 4.67-.83-1.86 5.39-3.51 10.88-6.79 15.55-3.06 4.36-6.95 7.12-12.05 7.36-11.74.56-21.13-4-27.15-15.59-2.38-4.58-4.45-9.28-5-14.58 1.45-.49 2-2 2.86-3.11 2 2.14 2.11 5.19 3.1 7.81 3.51 9.29 8.37 16.94 18.11 19.32 9.27 2.26 15.53-1.09 19.36-10.54.77-1.91.61-4.46 2.89-5.39Z"
-      style={{
-        fill: "#171717",
-      }}
-    />
-    <path
-      className="d"
-      d="M100.1 165.81c1.44.79 2.25-2.38 3.84-.69-2.79 7.08-7.22 12.27-14.61 12.44-7.89.19-14.79-2.8-19.22-10.64-3.69-6.53-5.42-13.86-7.2-21.16 3.3 0 1.32-4 3.14-5.12 1.86 2.32 1.85 5.33 2.5 8.07 1.08 4.61 2.14 9.22 4.23 13.44 3 6 7.56 9.28 13.68 10.12a10.22 10.22 0 0 0 10-3.89c.94-1.16 1.71-2.93 3.64-2.57Z"
-    />
-    <path
-      d="M164.12 167.36c-1.25 3.95-2.26 8.06-4.7 11.39s-4.85 5.33-7.72 6.27c-3.71 1.22-8.11.55-13.93-1.87-11-4.58-14.93-15.34-17.16-27a19.25 19.25 0 0 0 0-2.52c-.75-6.06.72-8.18 6.06-9.46a34.87 34.87 0 0 1 9.93-.81c2 .1 3.9.32 5.85.61s3.76.63 5.65 1a78.62 78.62 0 0 1 9.4 2.32c7.69 2.5 9.72 6.62 7.73 15.2-.4 1.62-.75 3.25-1.11 4.87Z"
-      style={{
-        fill: "#2e2e2e",
-      }}
-    />
-    <path
-      d="M100.1 165.81c-9 12-22.78 9.91-28.95-4.39a68.9 68.9 0 0 1-5.1-20.78c10.68-3.48 21.33-6.72 32.12 0 5.35 3.35 7 6.82 5.51 13.29-.94 4.07-2.37 7.94-3.58 11.88Z"
-      style={{
-        fill: "#2d2d2d",
-      }}
-    />
-    <path
-      className="i"
-      d="m137.31 143.61 22.89 35.14s-3.3 8.27-10.41 6.72c-2.19-.48-22.86-41.75-22.86-41.75a27.7 27.7 0 0 1 10.38-.11ZM91.55 137.7l11.44 18.91-3.19 8.91-13.86-28.75 5.61.93"
-    />
-    <path
-      className="j"
-      d="M161.38 35.12c-.25-2.23 0-4.31 1.8-5.93 1.06-.3 1.95-1 3-1.35a3.41 3.41 0 0 1 3.59.7 3.19 3.19 0 0 1 .76 3.48A18.32 18.32 0 0 1 168 37.4c-1 1.46-.87 2.08.83 2.42a112.94 112.94 0 0 1 12.39 3.67 60.2 60.2 0 0 1 23.37 14.78c1.55 1.57 2.63 3.49 4.13 5.09a38 38 0 0 1 5.5 16.58 5.29 5.29 0 0 1-.13 2 28.12 28.12 0 0 1-3.45 13.82c-2.14 3.86-5 7.13-9.55 8.23a7.81 7.81 0 0 1-6.85-1.29 25.24 25.24 0 0 1-7-8.76 81.09 81.09 0 0 0-14.85-19.71 53.37 53.37 0 0 0-17.15-11c-11.09-4.59-22.76-5.09-34.5-5.46-5.93-.19-11.88-.29-17.78-1-2-.23-2.69-1.31-2.13-3.26.78-2.65 2.73-4.5 4.64-6.26 4.28-4 9.48-6.33 15-8a65.45 65.45 0 0 1 16.11-2.82 105.67 105.67 0 0 1 13.55.21c3.15.26 6.26.73 9.37 1.2 2 .3 1.78.21 2.29-1.87.08-.33-.09-.7.06-1a2.09 2.09 0 0 1 2-.36c.71.26 1.21-.19 1.51-.79s.87-1.48.47-2.35c-.24-.53-.73-.51-1.2-.29a1.76 1.76 0 0 0-1.08 1.76 1.61 1.61 0 0 1-1.19 1.77c-.43-.07-.59.42-.98.41Z"
-    />
-    <path
-      className="k"
-      d="m161.38 35.12.86-.52c0 .11 0 .18-.05.22s-.11 0-.14 0c.24 1 .43 2 .72 3s0 1.38-1.06 1.22c-1.46-.23-2.92-.43-4.36-.68a85.16 85.16 0 0 0-25.44-.78c-8.68 1.06-17 3.36-24.19 8.69-2.57 1.92-4.94 4.14-6.14 7.25-.66 1.71-.23 2.28 1.64 2.46 8.21.79 16.46.73 24.69 1.16 8.52.45 17 1.29 25.09 4.25a57 57 0 0 1 25.11 17.87 99.73 99.73 0 0 1 11.31 17 16.25 16.25 0 0 0 6.39 6.4c2.46 1.45 5 .79 7.39-.42 3.59-1.86 5.92-5 7.74-8.48a25.45 25.45 0 0 0 2.8-10.76 5.18 5.18 0 0 1 .3-1 2.34 2.34 0 0 1 .22.73c-.36 8.58-3 16.19-9.69 22-4.65 4-10.07 4.71-14.89 2.11a16.65 16.65 0 0 1-1.79-1.27c-4.32-1.64-5.94-5.62-8-9.23a82.72 82.72 0 0 0-12.73-17.24 51.46 51.46 0 0 0-13.68-10 58.66 58.66 0 0 0-21.94-5.88c-2.72-.21-5.38-.9-8.11-1.09-5.74-.83-11.54-.44-17.3-.89-2.67-.21-5.35-.37-8-.71-3.67-.49-5.28-3-4.3-6.54 1-3.77 3.59-6.5 6.52-8.87 7.2-5.82 15.66-8.64 24.65-10.1 11.72-1.9 23.37-1.16 35 1.12 1 .24 1.64.32 1.38-1.02Z"
-    />
-    <path
-      className="k"
-      d="M123.43 62.17c1.28-.48 2.61-.14 3.88-.06 7.14.45 14.22 1.38 21 3.94a54 54 0 0 1 19.39 12.49 83.42 83.42 0 0 1 14.15 19.56 23.8 23.8 0 0 0 6 7.36c-1.07-1.09-2-.74-3.18 0a7.4 7.4 0 0 1-9.65-1.79c-4.67-6-9.31-11.94-14.39-17.56-7-7.72-14.43-14.87-24-19.42a60.12 60.12 0 0 0-11.44-4c-.57-.12-1.19-.33-1.76-.52Z"
-    />
-    <path
-      d="M208.67 63.36a1.48 1.48 0 0 1-1-.72C201.89 54.79 194 49.75 185.2 46a103.12 103.12 0 0 0-18.75-5.83c-.39-.08-.94 0-1.09-.46s.42-.65.67-1a20 20 0 0 0 3.73-6.87 2.53 2.53 0 0 0-.51-2.78 3.86 3.86 0 0 0-3.23-.79c-1 .17-1.7 1.24-2.84.9a8.45 8.45 0 0 1 5.89-2.17c2.82.08 4.29 2 3.67 4.77a17.2 17.2 0 0 1-2.13 5.23c-.71 1.15-.53 1.62.8 2a84.85 84.85 0 0 1 16.42 6.37c8.11 4.17 15.23 9.51 20.4 17.18.17.24.29.54.44.81Z"
-      style={{
-        fill: "#020202",
-      }}
-    />
-    <path
-      className="k"
-      d="M207.25 94.75a17.29 17.29 0 0 1-2.62 2.72c-3.34 2.28-6.52 1.68-8.68-1.8-2-3.17-3.8-6.42-5.7-9.63a70.71 70.71 0 0 0-12.65-15.57 58.3 58.3 0 0 0-10.85-8.29 61.49 61.49 0 0 0-12.62-5.6c-2.88-.86-5.77-1.54-8.7-2.12a91.14 91.14 0 0 0-14.84-1.59c-4.45-.15-8.9-.47-13.36-.55-2.1 0-4.2-.18-6.28-.41-.41 0-.92 0-1-.52s.3-.63.54-.89c2-2.07 4.46-3.33 6.91-4.63 5.76-3 12.1-4.17 18.45-4.94a74.23 74.23 0 0 1 12.74-.09 109.75 109.75 0 0 1 31.38 6.25 83.52 83.52 0 0 1 9.18 3.65c7.12 3.51 13.51 8 18.21 14.67 2.77 3.9 3.72 8.31 4.51 12.89.83 4.77-.84 9-2.25 13.29-.47 1.29-1.62 2.12-2.37 3.16ZM162.05 34.84l.19-.24c.08-.39.44-.64.34-1.15a2.68 2.68 0 0 1 1.84-3.24c.61-.25 1.16-.52 1.75-.05a1.8 1.8 0 0 1 .73 1.8 7 7 0 0 1-1.12 2.74c-.7 1-1.55 1.59-2.73.53a3.21 3.21 0 0 0-1-.39Z"
-    />
-    <path
-      d="M110.24 301.24c-1-1.75-.12-3.31-1.67-4-.91 4.42-2 9-4.22 13-1.3-.44.17-2.17-.7-2.91-1.61 2.91-3.15 5.9-4.89 8.71-1.05 1.71-2.12 2.17-3-.31a64.52 64.52 0 0 1-2.59-11.17c-.56 2.47-.81 3.54-1.33 5.86-1.07-5.49-1.42-9.75-.57-14.35-1.43.65-1.11 3.05-2.82 3.31.44-7.38-.22-15.38 6.55-19.59 4.56-2.83 9.42-3 13.5 1 5.67 5.38 2.76 12.71 1.74 20.45Z"
-      style={{
-        fill: "#0d0d0d",
-      }}
-    />
-    <path
-      className="n"
-      d="M106.05 281.27 97.66 313l5.59-10.26 1.87 1.86 2.8-10.26h1.86s3.74-9.34-3.73-13.07Z"
-    />
-    <ellipse className="n" cx={74.33} cy={329.31} rx={5.6} ry={7} />
-    <path
-      className="j"
-      d="m79.93 311.12 1.87-27.05s-36.39 58.77-29.86 74.63 101.69-21.46 101.69-21.46L148 327s-66.2 25.17-83 19.57Z"
-    />
-    <path
-      className="n"
-      d="m143.37 314.85.93-21.45s31.93 52.14 20.53 68.1c-4.67 6.53-98-19.59-98-19.59l4.66-12.13s74.64 21.46 80.24 17.73Z"
-    />
-    <path
-      className="k"
-      d="M140.19 320.38c6.17-6.62 16.56-4.63 20.1.75s-6.82 8.07-10.39 16.11c-3.73 8.4-11.25-1.9-11.47-2.23s-4.43-8.01 1.76-14.63Z"
-    />
-    <path className="n" d="m74.45 138 5.07-1.28 15 34.23s-4.05 3.53-9 3.08Z" />
-    <path
-      className="o"
-      d="M116.69 277a52.91 52.91 0 0 1-15.69-3.23c-1.77-.65-3.47-1.48-5.2-2.22-15.93-8.8-23-23-24.64-40.48-.08-.79-.06-1.56-1.15-1.76-.54-.1-.66-.76-.71-1.27s.5-.53.83-.65a11.71 11.71 0 0 1 5.72-.54 2.67 2.67 0 0 1 .93.11c2 .39 2.33.82 1.81 2.76a4.13 4.13 0 0 0 0 3.18.61.61 0 0 1-.08.56c-.92.39-1.39-.54-2.28-.79a13.92 13.92 0 0 0 1.17 5 19.19 19.19 0 0 0 4.29 6.18 2.38 2.38 0 0 1 .84 1.11c-.46 1.21-1 0-1.52 0a4 4 0 0 1 1.08 3.19 3.38 3.38 0 0 0 .67 1.62c.14.26.56.47.39.79s-.63.2-1 .18a11.29 11.29 0 0 1-1.15-.27c1.41 5.43 6.58 11.26 11.77 14.5.3.18.61.36.9.57 3.9 2.89 8.4 4.45 13 5.88 1.73.54 3.49.95 5.27 1.33a40.58 40.58 0 0 0 10.94.71 17.69 17.69 0 0 0 2.9-.56 30.92 30.92 0 0 0 7.7-2.3.79.79 0 0 1 .79 0c.38.19-.18 1 .56.92.93-.15.82-.88.74-1.55a.87.87 0 0 1 .91-.65c1.09.84 1.09.85 2.37 0A15.84 15.84 0 0 1 142 268c.36-.14.6.07.5.51l-.88.82a32.57 32.57 0 0 0-3.46 2.19c-1 .08-1.56.81-2.34 1.2a36.88 36.88 0 0 1-18.43 4.28h-.7Z"
-    />
-    <path
-      className="p"
-      d="M115.6 243.83c3.46.63 7 1 10.44 1.36.62.06 1.23.17 1.85.26a6.23 6.23 0 0 0 2 .41c.49.09 1.1.07 1 .86a1.16 1.16 0 0 1-1.56.93c-.05.3.73.1.36.59a18.65 18.65 0 0 1-5.17-.27 169.94 169.94 0 0 1-17.93-3.2c-.63-.14-1.38-.11-1.72-.85.89-.38 1.7.34 2.46.23-.9-.05-1.86-.44-2.89-.44-.54 0-.83-.43-.71-1s.47-.63 1-.52c1.36.3 2.72.56 4 .84-.2-.13-.57 0-.75-.32l1.51.12a39.26 39.26 0 0 0 6.11 1Z"
-    />
-    <path
-      d="M78.2 233.45a.26.26 0 0 0 .15-.38c-.82-1.13-.41-2.22-.09-3.41.56-2 .4-2.18-1.6-2.61l-.8-.2a28.93 28.93 0 0 1 6.06 1c.22 0 .5.11.51.41s-.22.41-.43.51a5.12 5.12 0 0 1-1.29.58c-1 .21-.88.73-.49 1.42a13.51 13.51 0 0 0 3.24 3.42 5.21 5.21 0 0 1 .38 1.35c.06.67-.19 1-.87.87a18.51 18.51 0 0 1-4.28-2.22c-.26-.19-.56-.35-.49-.74Z"
-      style={{
-        fill: "#26100c",
-      }}
-    />
-    <path
-      className="o"
-      d="M83 236.14c.33 0 .74.18.59-.46a3.18 3.18 0 0 1-.11-1.5 58.19 58.19 0 0 0 7.48 4 18.77 18.77 0 0 0 2.62.89c.29 1 1.2 1.13 2 1.45.18.07.37.11.55.19.34.14.93.17.82.64s-.66.22-1 .22-.9-.17-1.34-.27c-2.11-.5-4-1.58-6.05-2.25q-2.35-.94-4.64-2.07c-.43-.2-.92-.26-.92-.84Z"
-    />
-    <path
-      className="o"
-      d="m96 241.26 1.08.06c-1.13-.38-2-.67-2.83-.94-.68-.23-1-.63-.64-1.33l6.19 1.79c1.28.5 2.7.48 4 1l4.26.87 1.68.45v.17l-4.6-1c-.45-.09-1-.32-1.09.44s.46.62.85.71a12.32 12.32 0 0 1 2.92.9l-2.86-.59a4.9 4.9 0 0 1-3.09-.38c-1.69-.51-3.41-.9-5.07-1.51-.44-.08-.8-.16-.8-.64Z"
-    />
-    <path
-      className="p"
-      d="M142.49 268.52c-.09-.12-.25-.36-.26-.35-1.15.65-2.54.74-3.6 1.61-1.23 1-1.25 1-2.18-.3 0-.43.32-.62.61-.8a28.5 28.5 0 0 0 8.14-7.6.73.73 0 0 1 .43-.25 4.49 4.49 0 0 1-1.59 0c-.35-.14-.58-.45-.27-.7.74-.59 1-1.53 1.88-2a2.29 2.29 0 0 0 1.33-2.6c.25-.76 1-1 1.55-1.39a34.39 34.39 0 0 0 7.27-6.85c-2.06.23-4 .85-5.92.74-.24-.13-.63-.43-.26-.55 1.28-.39 1.77-1.81 3-2.16l.52-.22c1-.74 2.3-.71 3.26-1.48.78-.61.15-1.08-.16-1.59-.36-.48-1.16-.72-.86-1.57a10.49 10.49 0 0 1 5.09 2.69 4.51 4.51 0 0 1 1.25 4c-.27 1.54-1 1.8-2.15.81-.24-.2-.47-.42-.78-.7a59.15 59.15 0 0 1-3.74 7.16A53 53 0 0 1 143.37 268c-.27.19-.49.5-.88.52Z"
-    />
-    <path
-      className="p"
-      d="M141.8 246.1a40.4 40.4 0 0 0 10.35-.8h.5c-1 .71-1.47 2.14-2.93 2.29v.11l.21.11c-.2.6-.74.49-1.17.53-2.26.22-4.53.39-6.81.47a9.81 9.81 0 0 1-2.09-.15.66.66 0 0 1-.4-1c.2-.5.3-1 .5-1.52.62-.48 1.23-.14 1.84-.04Z"
-    />
-    <path
-      className="p"
-      d="m129.29 245.66 7.26.34 3.52.2c.21.48 0 .86-.2 1.29-.11.25-.62.66 0 .92-.23.53-.72.38-1.13.4-1 .05-2.05-.06-3.07-.05s-2.12 0-3.18-.11c-.88-.31-1.94.15-2.71-.63l-.67-.2c0-.29.2-.15.29-.14.88.1 1.1-.69 1.21-1.17s-.62-.2-.91-.38c-.58-.35-1.35-.13-1.85-.68ZM135.81 272.73c.66-.63 1.19-1.51 2.34-1.2a6.14 6.14 0 0 1-2.34 1.2Z"
-    />
-    <path
-      d="m129.82 248 2.8.33c1.49.49 1.52.49 1.11 2.15a81.39 81.39 0 0 1-2.58 9.1 1.45 1.45 0 0 1-1.35.79 34.63 34.63 0 0 1-6.11-.09c-.41-.12-.87 0-1.15-.11a24.18 24.18 0 0 0 3.08.27 36.24 36.24 0 0 1 4.76.5c.86.37 1 1 .78 1.86a11.77 11.77 0 0 1-2.31 4.83 21.77 21.77 0 0 0-2.53 4.39 2.47 2.47 0 0 0-.26 1.13c-3.13 1.11-6.38.92-9.58.62a55.22 55.22 0 0 1-20.78-6.4 9.55 9.55 0 0 1-2.37-1.71c-.08-2 .17-4-.73-5.87-.48-1 .36-1.84.89-2.64a7 7 0 0 1 .91-.9 4.26 4.26 0 0 1 3.35-1.77 1.25 1.25 0 0 0 1.1 1.05c1.15.24 2.27.6 3.51 1-.84-.39-1.77-.25-2.32-1-.14-.21.57-.09.16-.37a10.46 10.46 0 0 1-1.27-1.22c-.64-.63-.44-1.1.28-1.47a6.9 6.9 0 0 1 2.05-.79h-.11c-3 .45-3.06.31-2.47-2.65a22.5 22.5 0 0 0 .52-2.3c0-.15-.25-.06-.09-.07s.06 0 0 .1a30.43 30.43 0 0 1-1.64 4.91c-.18.45-.52.58-.93.24-.63-.52-1.38-1-1.48-1.9a38.6 38.6 0 0 1 1-7.41 1.56 1.56 0 0 1 .56-1c2.75.69 5.41 1.69 8.2 2.21a5.15 5.15 0 0 0 .54.23 154 154 0 0 0 24 4 2.38 2.38 0 0 0 .46-.04Z"
-      style={{
-        fill: "#dfdfdf",
-      }}
-    />
-    <path
-      d="M126.06 273.17c-.46-.16-.48-.47-.3-.85a38.87 38.87 0 0 1 3.74-6.22 9.54 9.54 0 0 0 1.48-4c.09-.49-.43-.51-.55-.83.37-.42.85-.34 1.33-.33a96.82 96.82 0 0 0 10.77-.37c.56-.06 1.06.23 1.61.13l1.71-.07a30 30 0 0 1-9.4 8.89l-.89.51a1.39 1.39 0 0 1-1.29.66l-3.38 1.31a26.81 26.81 0 0 1-4.83 1.17Z"
-      style={{
-        fill: "#f0f0f0",
-      }}
-    />
-    <path
-      d="M94.57 256.46a9.42 9.42 0 0 0-1.23 1.39 2.51 2.51 0 0 0-.09 2.77c.75 1.43.19 3.23.55 4.84 0 .24-.29.19-.47.22-6-3.7-10.52-8.54-12.82-15.51l1.65.34a39.54 39.54 0 0 1 6.84 2 59.44 59.44 0 0 1 6 2c.72.84-.45 1.26-.43 1.95Z"
-      style={{
-        fill: "#f4f3f3",
-      }}
-    />
-    <path
-      d="M83 236.14c1.75 1.08 3.73 1.67 5.55 2.61.63 1 0 1.91-.42 2.74a19.54 19.54 0 0 0-1.61 3.76 2.66 2.66 0 0 1-1.46 1.75c-1.34 0-2-1-2.68-2a22.68 22.68 0 0 1-5.26-7.2 14.35 14.35 0 0 1-1.22-5.8c.89.43 1.42 1.24 2.3 1.5a23.65 23.65 0 0 0 4.8 2.64Z"
-      style={{
-        fill: "#fefefe",
-      }}
-    />
-    <path
-      d="M82.38 245a29.44 29.44 0 0 0 2.51 2c1.75.77 3.16 2.09 4.85 3 1.25.72 2.53 1.39 3.79 2.06.39.21 1.11.46.29 1.05-.69 0-1.37-.06-2.06-.07-.41-.17-.9 0-1.15-.13a15.71 15.71 0 0 0 2 .16c.82.16 1.65-.08 2.46.1.43.1.91.22.92.77a.83.83 0 0 1-.85.93l-12-3.86c-.33-.11-.64-.28-1-.42l.73.09c-.61-.95-1.33-1.84-1.14-3.09a2.34 2.34 0 0 0-.85-2.1c-.32-.3-.6-.64-.17-1s.45.26.68.39.51.85.99.12Z"
-      style={{
-        fill: "#2d110e",
-      }}
-    />
-    <path
-      className="p"
-      d="m134.27 270.65 1.29-.66a.9.9 0 0 1 .23 1.33 1.29 1.29 0 0 1-1.41.61c-1.04-.27.16-.87-.11-1.28Z"
-    />
-    <path
-      className="w"
-      d="M89.57 250.09a22 22 0 0 1-4.68-3.09 4.41 4.41 0 0 0 1.56-2.7 8.08 8.08 0 0 1 1.24-2.64 6.24 6.24 0 0 0 .84-2.87c1.48.57 3 1.07 4.42 1.73.55.25 1.14.25 1.66.51a16.62 16.62 0 0 0-2.88 4.64 4.29 4.29 0 0 0-.39 2.05c.12.62 0 1.2-.68 1.3s-.46.98-1.09 1.07Z"
-    />
-    <path
-      className="x"
-      d="M89.57 250.09c0-.75.16-1.29 1-1.38.42-.05.56-.29.51-.75-.27-2.22.83-4 2.11-5.61.37-.47.47-1.39 1.39-1.36l1.34.27.71.37a65.32 65.32 0 0 0-1.31 8.69.52.52 0 0 1-.44-.07 1.21 1.21 0 0 1-.44-.37c.42.65.91 1.25 1.38 1.87.64.85.46 1.26-.6 1.32-.47 0-1 .24-1.43-.18.42-.35.06-.39-.18-.52a29 29 0 0 1-4.04-2.28Z"
-    />
-    <path
-      d="M142 248.51c2.64-.27 5.32-.18 7.94-.71a57.6 57.6 0 0 0 6.55-1 36 36 0 0 1-9.37 8.62 7.09 7.09 0 0 1-4.74 2.13 1.53 1.53 0 0 1-1-1.59c0-1.22-.09-2.44 0-3.66a3 3 0 0 0-.17-1.39c-.38-1.18-.21-1.59.79-2.4Z"
-      style={{
-        fill: "#fdfefd",
-      }}
-    />
-    <path
-      d="M142.54 257.41a22.52 22.52 0 0 0 4.59-2c.46 1.25.2 2.15-1.21 2.89-1 .49-1.07 1.67-2 2.06-.09 0 .13.23.27.29a78.43 78.43 0 0 1-12.54.55h-1.17l-5-.52c-.62-.14-1.37.09-1.82-.6h6a4.71 4.71 0 0 0 1.44-.38c1-.55 2.2-.35 3.3-.58.42-.09.85-.1 1.27-.18 2.33-.41 4.58-1 6.87-1.53Z"
-      style={{
-        fill: "#3c1410",
-      }}
-    />
-    <path
-      className="w"
-      d="M142.54 257.41a18.1 18.1 0 0 1-5.24 1.51c-.52.1-1 .17-1.56.26-.62-.34-.14-.69.09-.82 1-.64.84-1.6.74-2.54-.27-2.44-.74-4.86-.84-7.32l4.18-.09 2.1.1c-.57.87-1 1.66-.25 2.76.3.43.11 1.3 0 1.93-.21 1 .16 1.94 0 2.92a1.25 1.25 0 0 0 .78 1.29Z"
-    />
-    <path
-      className="x"
-      d="M135.73 248.5a1.38 1.38 0 0 1 .41 1.58c-.07.18-.17.44-.09.57.89 1.5.26 3.24.77 4.78a10.51 10.51 0 0 1 .15 2.09.74.74 0 0 1-.56.79c-.39.14-.84.28-.67.87a19.28 19.28 0 0 1-4.67.52 1.32 1.32 0 0 1 .14-1.63 4.46 4.46 0 0 0 .8-2.18c.26-1.91 1-3.69 1.45-5.56.25-1 .17-1.63-.84-2a17 17 0 0 1 3.11.17Z"
-    />
-    <path
-      d="M93.82 252.89c.63.29 1.24-.07 1.84-.14s.21-.57 0-.8a7.78 7.78 0 0 1-1.78-2.76c.71.23 1 .86 1.46 1.13a9.29 9.29 0 0 0 1.77 1.7l2.16-6.02c.3 1.6-.27 3-.56 4.45-.21 1-.07 1.41 1.06 1.18a10.06 10.06 0 0 1 2.47-.13c-1 .44-1.94.92-2.93 1.31-.64.25-.78.45-.21.94a11 11 0 0 0 1.86 1.66c-.36.27-.61 0-.86.07a2.26 2.26 0 0 1-2.42-1 1.52 1.52 0 0 1-1.24.53c-.8 0-1.23 1-1.87 1.48-.3-.09-.34-.24-.15-.5s.65-.67.7-1.17c.33-.27.82-.55.71-1s-.72-.34-1.12-.41c-.78-.12-1.56-.19-2.34-.28-.25-.05-.68.23-.63-.36Z"
-      style={{
-        fill: "#fcfbfb",
-      }}
-    />
-    <path
-      d="m97.68 254.45 2.42 1a18.6 18.6 0 0 0 2.73 1.19c-1.84-.38-3.37-.67-4.89-1-.72-.22-.39-.76-.26-1.19Z"
-      style={{
-        fill: "#462524",
-      }}
-    />
-    <path
-      d="M123.66 260.11c.5.53 1.26.26 1.82.6A8.49 8.49 0 0 1 122 260Z"
-      style={{
-        fill: "#715355",
-      }}
-    />
-    <path
-      d="m91.74 252.76.63.36a5.26 5.26 0 0 1-2.5-.35 5.54 5.54 0 0 1 1.87-.01Z"
-      style={{
-        fill: "#a48582",
-      }}
-    />
-  </svg>
-);
-
-export default SvgFries;
