--- client/src/components/pages/SvgPotato 2.js
@@ -1,1671 +1,0 @@
-import * as React from "react";
-
-const SvgPotato = (props) => (
-  <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 249 288" {...props}>
-    <defs>
-      <style>
-        {
-          ".m{fill:#333}.r{fill:#404040}.v{fill:#161616}.aa{fill:#4d4d4d}.ab{fill:#e6a930}.ad{fill:#edc367}.af{fill:#f2d28a}.ag{fill:#eac97d}.ak{fill:#ebca87}.ap{fill:#e5a734}.aq{fill:#eecf8b}.as{fill:#edcd88}.at{fill:#e5a931}.av{fill:#efd08d}.aw{fill:#eecd86}.bc{fill:#c08017}.bg{fill:#c98c22}.bk{fill:#e8c37c}.br{fill:#cb9234}.bs{fill:#f1d48f}.bu{fill:#dea12d}.bv{fill:#f4d68c}.bx{fill:#9b5f0d}.cx{fill:#c9871e}.cy{fill:#f0c874}.cz{fill:#9e6310}.db{fill:#ebcc92}.dg{fill:#f2d48c}.dh{fill:#f1d28a}.do{fill:#fefefe}.eb{fill:#af791e}.ej{fill:#f4d890}.fg{fill:#efd18a}.fq{fill:#f5d995}.ga{fill:#e3a82f}.gj{fill:#b27f41}.hs{fill:#a86107}.in{fill:#eecc8b}.iu{fill:#8c4e08}"
-        }
-      </style>
-    </defs>
-    <path
-      d="M95.08 226.82c-8.52-11.46-13.26-24.71-15.21-38.88-3.78-27.48-3.68-54.72 1.35-81.77 4.82-26 17.45-43.74 41.12-50.58 7.81-2.26 16-3 23.88 1.57 7.53 8 11.26 17.31 11.32 28.78.06 9.08.44 18.41 4.41 27 2.87 6.2 6.13 12.18 9 18.4 7.27 15.8 12.19 31.92 11.17 49.76a56 56 0 0 1-10.61 30c-9.9 13.83-23.15 21.6-39.86 22.14-10.07.32-20.07 1.22-29.95-2.93a34.59 34.59 0 0 1-6.62-3.49Z"
-      style={{
-        fill: "#cd9f4c",
-      }}
-    />
-    <path
-      d="M95.08 226.82c6.85.93 13.1 4.72 20.19 5.09 14.75.79 29.78.73 41.82-8.26 13.36-10 22.68-23.78 23.55-43.52.83-18.69-4.21-35.46-12.76-51.18A98.81 98.81 0 0 1 155.53 80c.06-8.56-4.16-14.76-9.34-20.29-1.59-1.69-1-1.76 0-2.57C157 61 163.7 69.72 165.64 81.2c3.5 20.7 13.15 38.44 21.74 56.82 9.34 20 14.49 40.78 8.11 62.46-7.81 26.58-22.75 45.36-49 50.83A53.3 53.3 0 0 0 132 252c-16.31-1-27.78-10.93-36.92-25.18Z"
-      style={{
-        fill: "#c08b40",
-      }}
-    />
-    <path
-      d="M132 252a23.5 23.5 0 0 1 14.44-.73A57.18 57.18 0 0 1 132 252Z"
-      style={{
-        fill: "#b18522",
-      }}
-    />
-    <path
-      d="M94.49 178.44a23.17 23.17 0 0 0-3 .19c-4.2.79-5.29-1.54-5.81-5.78-2.31-18.88-2.44-37.69 1-56.06 3.14-16.63 7.86-32.72 20.12-44.28A30 30 0 0 1 117.47 66c3.52-1.26 5.76 1.33 6.79 4.64 4.08 13.13 2.4 24-7.18 33.82-9 9.19-12.68 21.77-13.32 35.87-.58 12.71 3.29 25.47.33 38.05a27.35 27.35 0 0 1-6.53 13c-3.29 3.52-5 1.38-7-1.41-.3-.42-.27-1.11-.47-1.64-2.95-7.96-2.95-7.95 4.4-9.89Z"
-      style={{
-        fill: "#e6c168",
-      }}
-    />
-    <path
-      d="M161 195.27a2.07 2.07 0 0 1-2.11-1.86 1.9 1.9 0 0 1 1.64-2.28c1.24-.26 1.89.52 2 1.82s-.11 2.21-1.53 2.32Z"
-      style={{
-        fill: "#a5772d",
-      }}
-    />
-    <path
-      d="M113.66 198.26c-.24 1.24-.86 2.1-2.06 1.85a1.87 1.87 0 0 1-1.52-2c.08-1.29.84-2.06 2-1.82a2 2 0 0 1 1.58 1.97Z"
-      style={{
-        fill: "#9e7026",
-      }}
-    />
-    <path
-      d="M129.05 71.59c2.26-1.45 4.2-1 6-1.13 1.29-.1 2.67-.66 4.64.85-2.46.13-4.28.27-6.1.32-1.3.05-2.59-.01-4.54-.04Z"
-      style={{
-        fill: "#b28435",
-      }}
-    />
-    <path
-      d="M159.59 241.51c-2.71.11-3.65-1.61-3.75-4.27s1.22-3.66 3.43-3.7c2.73 0 3.58 2 3.82 4.41.28 2.71-1.36 3.51-3.5 3.56Z"
-      style={{
-        fill: "#9c6b2a",
-      }}
-    />
-    <path
-      d="M193.12 170.78c0 1.3-.41 2.17-1.7 2.05a2.42 2.42 0 0 1-2.25-2.54 1.53 1.53 0 0 1 1.58-1.68c1.42-.13 2.25.55 2.37 2.17Z"
-      style={{
-        fill: "#a16b29",
-      }}
-    />
-    <path
-      d="M150.56 235.33c1.27-.18 1.91.53 2.05 1.85s-.11 2.27-1.45 2.32a2.06 2.06 0 0 1-2.24-1.83c-.22-1.33.61-1.97 1.64-2.34Z"
-      style={{
-        fill: "#9f6d29",
-      }}
-    />
-    <path
-      d="M171.21 118.06a2 2 0 0 1 1.63 1.46c.36 1.26-.49 1.62-1.41 1.69s-1.8-.14-2-1.31.28-1.78 1.78-1.84Z"
-      style={{
-        fill: "#a06d26",
-      }}
-    />
-    <path
-      d="M90.37 156.24a1.73 1.73 0 0 1-1.65 1.42 1.76 1.76 0 0 1-1.88-2.06 1.94 1.94 0 0 1 2-1.94c.97 0 1.59.6 1.53 2.58Z"
-      style={{
-        fill: "#d0a851",
-      }}
-    />
-    <path
-      className="m"
-      d="M128.93 22.56s12.21 1.77 13 2.23 7.4 3.51 7.4 3.51l5.33 5.52.07 3.29-26.88-2.22-28.58-4.62-6 .12 2.56-2.17 13.52-4.41L127 21.86Z"
-    />
-    <path
-      d="m87.31 39.06-.78 10.62-.39 5.32c0 .15.11.42-.16.44a10.14 10.14 0 0 0-3.36 1.28A44.06 44.06 0 0 0 74.81 61a16.41 16.41 0 0 0-4 3.79 8 8 0 0 0-.5 8.71 24.56 24.56 0 0 0 6.54 7.73 87.42 87.42 0 0 0 15.47 10.13 183.33 183.33 0 0 0 22.48 10.24 116.37 116.37 0 0 0 27.8 7.06 68.47 68.47 0 0 0 18.14-.22c.49-.07 1-.2 1.46-.31a2.24 2.24 0 0 0 .42-.45 1.66 1.66 0 0 1 1.46-.83 2.33 2.33 0 0 0 1.8-.78 1.91 1.91 0 0 1 2.1-.33 4 4 0 0 0 1.44.49 34 34 0 0 0 15-9.89 14.33 14.33 0 0 0 2.28-4.78 20.62 20.62 0 0 0 1.21-8.5 20.06 20.06 0 0 0-5.67-12.3 41.3 41.3 0 0 0-15.14-9.87 6.3 6.3 0 0 0-2.39-.74c-.17 0-.23-.12-.28-.25l-.15-.43c-2.65-7.74-5.39-15.46-7.94-23.23A18.92 18.92 0 0 0 148.05 26a22.3 22.3 0 0 0-5.35-2.54 2.27 2.27 0 0 0-.78 1.31c-.12.34.15.5.44.64 1.27.65 2.57 1.23 3.8 2a23.42 23.42 0 0 1 7.48 6.7 2.75 2.75 0 0 1 .19.28c.29.49.25.58-.34.71a25.62 25.62 0 0 1-4.54.48 48.5 48.5 0 0 1-13.35-1.39c-4.89-1.12-9.8-2.2-14.76-3.07a124.82 124.82 0 0 0-14.2-1.73 58.72 58.72 0 0 0-9.47 0c-.71.08-1.42.19-2.13.27a7.36 7.36 0 0 0-1.84.55c-.2.05-.1-.12-.12 0s0 0 .06 0c1.62-.49 3.18-1.11 4.75-1.72 1.36-.53 2.75-.94 4.12-1.4a93.24 93.24 0 0 1 16.4-3.84 55.86 55.86 0 0 1 9.75-.56 3.43 3.43 0 0 0 .64 0c.23-.05.37-.2.27-.38-.22-.41-.18-.91-.52-1.26a70.72 70.72 0 0 0-14.93 1.27 105.9 105.9 0 0 0-15.9 4.06 29 29 0 0 0-8.06 4.07c-.17.13-.32.27-.49.39a3.41 3.41 0 0 0-1.53 3c0 1.45-.16 2.92-.25 4.38a.33.33 0 0 0 .21.31c.36.18.77.51 1.15.14s0-.77-.26-1.07c-.68-.68-.75-1-.33-1.52a1 1 0 0 1 1.17-.39c.44.18 1 .26 1.26.79-.48 0-.76-.24-1.1-.35s-.59-.25-.85 0a.65.65 0 0 0 0 .8c.15.28 0 .18-.24.42-.61.76.25 1.3-.2 1.7-.19.16.53.36.35.3-.39-.15-.76-.43-1.24-.29Z"
-      style={{
-        fill: "#010101",
-      }}
-    />
-    <path
-      d="M137 23.65a17.53 17.53 0 0 1 3.53 1c.56.21.91.2 1.14-.4a7 7 0 0 1 .48-.95 21.67 21.67 0 0 0-4-1.08 4.68 4.68 0 0 1-2.14.05 14.79 14.79 0 0 1-1.3-.71 43.32 43.32 0 0 0-5.11-.47c-.65.31-.34.64 0 .91a7.25 7.25 0 0 0 1 .61l.08.06a1 1 0 0 0 .87-.14 2.25 2.25 0 0 1 2.09-.09 20.15 20.15 0 0 1 2.05 1 1.93 1.93 0 0 0 1.31.21Z"
-      style={{
-        fill: "#030303",
-      }}
-    />
-    <path d="M89.13 37.22a1.61 1.61 0 0 1-.35-.39c-.11-.21-.24-.45 0-.67s.42 0 .62.07a8.62 8.62 0 0 0 1 .39c.14.06.33.16.42-.06s.16-.41-.08-.52a6.45 6.45 0 0 0-1.59-.66 1.17 1.17 0 0 0-1.2.67 1 1 0 0 0 .13 1.3c.22.23.46.43.67.65a.36.36 0 0 1 0 .54.38.38 0 0 1-.51 0c-.27-.13-.53-.27-.79-.41a2.75 2.75 0 0 0-.08.9 5.41 5.41 0 0 1 1.17.42" />
-    <path
-      d="M169.42 106.23c-.5-.24-1-.48-1.5-.7a2 2 0 0 0-2.25.48 1.58 1.58 0 0 1-1.32.64 2 2 0 0 0-2.15 1.48 2.45 2.45 0 0 0 1-.17 1.69 1.69 0 0 1 1.33-.53 1.88 1.88 0 0 0 1.35-.49 1.67 1.67 0 0 1 2-.49 3.32 3.32 0 0 0 .55.09Z"
-      style={{
-        fill: "#3d3d3d",
-      }}
-    />
-    <path
-      d="M168.45 106.54c-1.26-.68-1.65-.65-2.66.28a2.12 2.12 0 0 1-1.43.46c-.49 0-1.08 0-1.16.68a33.92 33.92 0 0 0 5.25-1.42Z"
-      style={{
-        fill: "#060606",
-      }}
-    />
-    <path
-      className="r"
-      d="M134.76 21.52a2.69 2.69 0 0 0 3.44.66 19.73 19.73 0 0 0-3.44-.66Z"
-    />
-    <path
-      d="M161 72.38a100.56 100.56 0 0 1-15.71 4.76 87.55 87.55 0 0 1-34.64 1 69.3 69.3 0 0 1-25-9.76l-.38-.2c-.11 1.41-.21 2.75-.31 4.09-.15 2-.28 4.1-.44 6.15a.81.81 0 0 0 .43.85A68.19 68.19 0 0 0 96 84.88L98.82 86a32.9 32.9 0 0 0 4.78 1.65c.74.19 1.47.43 2.23.6a1.9 1.9 0 0 0 .75.07l4.23.89a75.85 75.85 0 0 0 11.28 1.34A78 78 0 0 0 137 89.8a83 83 0 0 0 18.64-5A102.05 102.05 0 0 0 170 77.66c.36-.21.42-.39.3-.74-.81-2.31-1.6-4.63-2.39-7-.11-.35-.16-.55-.69-.29-1.2.61-2.46 1.11-3.69 1.66-.41.31-.93.18-1.37.36a4.48 4.48 0 0 0-.55.21c-.24.14-.38.41-.61.52Z"
-      style={{
-        fill: "#fff",
-      }}
-    />
-    <path
-      d="M106.58 88.3c-2.6-.74-5.2-1.46-7.76-2.32.55.57 1.39.64 2.07 1 1.22.63 2.51 1.14 3.78 1.68a1.45 1.45 0 0 0 1.91-.36Z"
-      style={{
-        fill: "#3f3f3f",
-      }}
-    />
-    <path
-      className="r"
-      d="M130.59 22.56c.5.3 1.09.13 1.63.22 1.18.21 2.39.32 3.59.47a15.51 15.51 0 0 0-2.37-1.11 2.1 2.1 0 0 0-1.84.14 2.16 2.16 0 0 1-1.01.28Z"
-    />
-    <path
-      d="M147.58 63.31a1.26 1.26 0 0 0 1.09.54c.3 0 .44-.44.45-.74s-.92-.56-1.15-.47a.59.59 0 0 0-.39.67Z"
-      style={{
-        fill: "#040404",
-      }}
-    />
-    <path
-      className="v"
-      d="M172.14 108.74c1.7.52 3.06-.24 4.49-1 1.71-1 3.56-1.69 5.35-2.52-4.55-1-9.09-2.11-13.67-3-2.63-.52-3.91 1.84-5.62 3.12s.65 2.52 0 3.95a4.25 4.25 0 0 0 1-.17c1.51-.84 1.94-3.49 4.05-3 1.62.27 2.62 2.17 4.4 2.62Z"
-    />
-    <path
-      d="M172.14 108.74c.24.27.78.77.46.87-4 1.33-7.34 4.67-11.84 4.67-2.51 0-5-.57-7.53-.82-9.5-1-18.37 1.71-27.1 4.9a5.15 5.15 0 0 1-4.61-.2c-4.43-2.23-8.92-4.33-13.41-6.45-.9-.43-1.93-1.65-2.83-.57-1.07 1.26-2.61 2.47-2.18 4.5a6.27 6.27 0 0 0 1.18 3c1.24-.68 1-2.29 2.26-3.36 2.4 5 7.48 5.72 12 7.06 4.27 1.29 7.07 3.65 8.65 7.68 1.11.08 2.29 1.48 3.31-.18.09-1.31-.54-2.46-.81-3.69-.62-2.86.11-5 2.76-6.05 6-2.34 12.24-4.1 18.82-3.49 4.6.42 5 .93 5.43 5.52.5 1.13.63 2.67 2.52 2.24.33-1.2.66-2.4 1-3.61.46-1.74 1.8-2.17 3.35-2.11s1.18 1.45 1.3 2.41c.31 2.52 2.13 4.23 3.3 6.28.93.17 1.91 1.49 2.75-.13-.25-2.27-1.6-4.16-2.11-6.34-.87-3.68-.27-5.42 3-7.32s6.61-3 10.35-2.21c2.84.58 5.89 0 8.52 1.6.91.71.66 3 2.77 2.06.29-.81.65-1.61.87-2.44 1.08-4.17.72-4.75-3.46-5.6-3-.6-5.91-1.18-8.86-1.76-3.67.13-6.46 2.67-9.86 3.54Z"
-      style={{
-        fill: "#1c1c1c",
-      }}
-    />
-    <path
-      d="M130.14 130.33c-1-.29-1.9-1-3-.34 1.35 3.06 2.57 6.17 4.78 8.77a10 10 0 0 0 7.85 3.89c7.44 0 13.24-2.89 16.67-9.76a22.51 22.51 0 0 0 2.73-8.55c-.93-.24-1.31-1.1-1.9-1.71-1.22 1.29-1.17 3.05-1.71 4.59-1.93 5.45-4.76 10-10.84 11.65-5.79 1.57-9.85-.17-12.57-5.51-.53-1.08-.54-2.56-2.01-3.03Z"
-      style={{
-        fill: "#171717",
-      }}
-    />
-    <path
-      className="v"
-      d="M170.56 127.6c-.89.49-1.5-1.31-2.45-.29 2 4 4.95 6.86 9.63 6.75 5-.12 9.26-2 11.81-6.69 2.13-3.87 3-8.14 3.88-12.4-2.09.07-1-2.26-2.15-2.86-1.1 1.39-1 3.12-1.32 4.72a33 33 0 0 1-2.25 7.87 10.7 10.7 0 0 1-8.33 6.23 6.86 6.86 0 0 1-6.44-2c-.63-.6-1.18-1.6-2.38-1.33Z"
-    />
-    <path
-      d="M130.14 130.33c.91 2.25 1.68 4.58 3.33 6.43a10.09 10.09 0 0 0 5.08 3.39c2.38.6 5.14.09 8.75-1.47 6.81-3 8.94-9.28 10-16.05a12 12 0 0 1-.07-1.46c.29-3.51-.71-4.69-4.12-5.27a23.9 23.9 0 0 0-6.31-.19c-1.23.12-2.45.3-3.67.53s-2.36.47-3.54.73a51.43 51.43 0 0 0-5.87 1.61c-4.78 1.66-5.93 4.09-4.41 9 .28.9.55 1.83.83 2.75Z"
-      style={{
-        fill: "#2e2e2e",
-      }}
-    />
-    <path
-      d="M170.56 127.6c6.07 6.63 14.71 5.05 18.16-3.36a36.85 36.85 0 0 0 2.56-12.13c-6.85-1.7-13.69-3.26-20.3.94-3.28 2.09-4.22 4.13-3.06 7.82.72 2.29 1.75 4.49 2.64 6.73Z"
-      style={{
-        fill: "#2d2d2d",
-      }}
-    />
-    <path
-      className="aa"
-      d="M146.34 115.88 133 136.79s2.34 4.67 6.79 3.57c1.37-.34 13.13-24.72 13.13-24.72a19 19 0 0 0-6.58.24Z"
-    />
-    <path className="m" d="m157.27 120.39-12.86 19.36s14.19-5.18 12.86-19.36" />
-    <path
-      className="aa"
-      d="m175.07 111.15-6.63 11.23 2.3 5.04 7.85-16.97-3.52.7"
-    />
-    <path d="M147.89 177.06c-.43-1.94-1.34-2.31-2.9-1.29a4 4 0 0 0-2.07 3.3 3.28 3.28 0 0 0 3.73 3.65 10 10 0 0 0 6.18-3.26 36.85 36.85 0 0 0 4.54-5.65 42.18 42.18 0 0 1 5.31-6.6c2.93-3 8.6-5.15 13.48-2.72.53.27.81.29 1.18-.24 3.11-4.44 9.41-5.09 13.44-3.8 3.86 1.24 7.27 3.4 10.93 5.05a16.45 16.45 0 0 0 7.31 1.9 8.42 8.42 0 0 0 2.54-.39c2.26-.8 3.05-2.45 2.23-4.65a4.2 4.2 0 0 0-4.88-2.34c-1 .33-1.14 1-.92 2a1.89 1.89 0 0 1-1.05-2 1.87 1.87 0 0 1 1.6-1.51c2.86-.76 7.61 1.46 8.78 4.15A5.5 5.5 0 0 1 217 168a19.14 19.14 0 0 1-9.81 8.05 47.65 47.65 0 0 1-19.18 3.2 15.92 15.92 0 0 1-7.6-2.18c-.25-.15-.51-.62-.85 0-2.15 3.89-5.82 5.89-9.67 7.61a48.77 48.77 0 0 1-14.18 4 19.72 19.72 0 0 1-12.79-2.54 5.59 5.59 0 0 1-2-8 11.19 11.19 0 0 1 3.53-3.71 3.07 3.07 0 0 1 .77-.38c.9-.32 1.83-.6 2.66.17.58.63.61 1.24.01 2.84ZM88.75 138.13l-11.78 3.18-25.52 48.26 4.91.91 25.52-42.79 6.87-2.28v-7.28z" />
-    <circle className="ab" cx={46.37} cy={197.78} r={5.13} />
-    <path d="m52.3 188.49-11.19 1.27-3.11 5.08 4.97 6.99 2.49-1.9-3.11-3.81 1.87-3.18 3.73.63-1.25 4.45 1.25 1.91 2.48-2.54 1.87-3.18 3.73-1.91.87-2.89" />
-    <path
-      style={{
-        fill: "#392726",
-      }}
-      d="m43.7 202.16 2.94 57.44h4.86l-2.46-57.44H43.7z"
-    />
-    <path className="ad" d="M47.03 259.6h-.93v5.4l7.02-.54-.47-4.86h-5.62z" />
-    <path
-      style={{
-        fill: "#cbcbcb",
-      }}
-      d="m52.61 187.37 1.23-2.31 4.37 2.31-1.85 3.11-3.75-3.11z"
-    />
-    <path
-      className="af"
-      d="M86.85 194.33c.07.42.39.62.61.9s.62.69.9 1.06a1.79 1.79 0 0 0 1.73 1 1.44 1.44 0 0 1 .93.38 15.68 15.68 0 0 1 1.98 1.27.25.25 0 0 1-.15.19c-.3 0-.51-.22-.75-.36a.8.8 0 0 0-.63-.14c-.62 0-1-.64-1.6-.77 1.44.93 2.91 1.79 4.31 2.79.22.24.59.18.78.48a.3.3 0 0 1 0 .35c-.52.15-.94-.15-1.39-.35-.08-.28-.32-.37-.51-.52A20.09 20.09 0 0 0 90.7 199c-.2-.14-.35-.12-.44.16a.32.32 0 0 1-.29.26c-.21 0-.33-.2-.43-.4-.23-.68-.8-1.06-1.13-1.66-.1-.18-.26-.32-.28-.55a2.84 2.84 0 0 0-1.57-1.81c-.25-.06-.38-.27-.53-.45a1.33 1.33 0 0 0-.82-.53.25.25 0 0 1-.2-.25c.1-.22.67 0 .39-.57-.52-.94-1.16-1.79-1.71-2.71-.19.16-.19.49-.47.44s-.27-.59-.38-.89c-.15-.48-.37-.93-.56-1.4a1.08 1.08 0 0 1-.15-.66c0-.06.11-.11.1-.14a5.11 5.11 0 0 0-.81-1.79s-.11 0-.16 0c-.3-.18-.35-.54-.49-.84a1.12 1.12 0 0 0-.62-.69.41.41 0 0 1-.22-.47 1.06 1.06 0 0 1 1.44-.12 1.68 1.68 0 0 1 .49 1.37c0 1.28.76 2.14 1.14 3.2a.29.29 0 0 0 .19.13c1.35.26 1.86 1.58 2.57 2.62a6.79 6.79 0 0 1 1 1.68v.93c0 .16 0 .42-.26.32a2 2 0 0 1-.16-.77.82.82 0 0 0 0-.49 18.48 18.48 0 0 0-1-1.7c-.08-.12-.2-.09-.31-.07-.49-.25-.62-.84-1-1.22-.16-.17-.25-.42-.59-.57a19.83 19.83 0 0 1 1.44 2 14.89 14.89 0 0 1 1.55 2.47c0 .13 0 .27.1.36s.14.36.32.14Z"
-    />
-    <path
-      className="ag"
-      d="M107.41 201.2a1.78 1.78 0 0 0 1.23-1.16c.25-.63.88-.86 1.35-1.24.65-.53 1.33-1 2-1.5a2.28 2.28 0 0 1 1.4-.46 1 1 0 0 0 .8-.4 16 16 0 0 1 1.12-1.21 2.2 2.2 0 0 0 .85-2c-.09-.71.42-1.31.66-1.95.18-.47.44-.9.61-1.38a2.51 2.51 0 0 1 1.73-1.78c.44-.11.41-.63.66-.9a.09.09 0 0 1 .15 0c.12.32-.08.56-.15.83a7.87 7.87 0 0 1-.49 1c-.08.18-.2.34-.4.13a.51.51 0 0 1 0-.32c-.27.1-.35.49-.64.52-.56.32-.66 1-.91 1.53-.1.2-.22.42 0 .64a2.44 2.44 0 0 1-.36 1.19 1.85 1.85 0 0 0-.25.76c.18-.25.15-.68.54-.73a.81.81 0 0 1-.14 1.07c-.37.42-.78.79-1.17 1.19a13.11 13.11 0 0 0-1.59 1.73 1.18 1.18 0 0 0-.34.64c-.12.35-.45.6-.41 1a.5.5 0 0 1-.35.49.25.25 0 0 1-.27-.29c0-.13.2-.2.18-.4-.28 0-.51.32-.8.14-.06-.49.4-.51.57-.79 0-.06-.1 0-.14 0s-.12-.05-.16 0a5 5 0 0 0-2.57 2c-.14.33-.44.45-.69.63a1.13 1.13 0 0 0-.49.8c-.31.39-.76.38-1.15.54a.55.55 0 0 1-.25 0 .18.18 0 0 1-.13-.32Z"
-    />
-    <path
-      d="M123.07 183.27a1.64 1.64 0 0 1 .78 1.58l-.06 2.7a1.78 1.78 0 0 1-.69 1.47 1 1 0 0 0-.39.79c0 .79-.13 1.58-.22 2.36a.55.55 0 0 0 .24.58 1.73 1.73 0 0 1 .73 1.56c0 1.2-.07 2.4-.11 3.6a1.3 1.3 0 0 1-.24.77c-.26-.41-.1-.86-.1-1.3a1.55 1.55 0 0 1 0-.21 19.91 19.91 0 0 0 .1-2.32c0-1.1-.25-1.6-1-1.91a2.27 2.27 0 0 0-.07 1.1.41.41 0 0 1 0 .19c-.12.17-.27.32-.47.21s-.18-.33-.12-.54c.17-1.17.25-2.36.35-3.54a11.23 11.23 0 0 1 .4-2.17c.07-.19.22-.21.37-.25a.26.26 0 0 1 .17.05.65.65 0 0 1 .17.33.56.56 0 0 1-.25.53 1.69 1.69 0 0 0 .72-1.5 22.2 22.2 0 0 0-.06-2.94l-.1-.09c-.12-.09-.25 0-.39 0a.42.42 0 0 1-.35-.1c-.07-.26.17-.43.17-.68s.23-.33.42-.27Z"
-      style={{
-        fill: "#e2a942",
-      }}
-    />
-    <path
-      d="M89.25 232.82c-.7-1.18-1.11-2.75-2.77-2.68-.09-.06-.08-.14 0-.23.28-.22.59-.06.87-.06.08 0 0-.13 0 0h-.06a.39.39 0 0 1-.27-.07 1.05 1.05 0 0 1-.42-1.21.42.42 0 0 1 .18-.31.38.38 0 0 1 .32.09 1.07 1.07 0 0 1 .18.19c1.18 1.31 2.06 2.9 3.17 4.26a1.8 1.8 0 0 1 .26.64c0 .15.09-.06.15-.05s.29.14.37.31a.45.45 0 0 1-.26.46c-.28.1-.56 0-.84 0a.29.29 0 0 1-.32-.21c-.09-.45-.41-.74-.56-1.13Z"
-      style={{
-        fill: "#d99c28",
-      }}
-    />
-    <path
-      d="M103.88 238.13a1.83 1.83 0 0 1-1.44 1.4 26.75 26.75 0 0 1-3.23.06 1.74 1.74 0 0 1-1.72-1.59c.1-.26.32-.42.49-.22a1.29 1.29 0 0 0 1.35.35 5.47 5.47 0 0 0 1.93.15 1.16 1.16 0 0 0 1.16-.35 2.54 2.54 0 0 1 .72.09c.14 0 .25.1.19.3a1.25 1.25 0 0 1-1 .68 20.07 20.07 0 0 1-2 .08c-.35-.27-.76-.18-1.15-.22h.75c.11.31-.1.3-.25.3a2.27 2.27 0 0 1-1.15-.15 10.44 10.44 0 0 0 2.9.11c.1 0 .19-.06.28 0a2.05 2.05 0 0 0 2-1 .23.23 0 0 1 .17.01Z"
-      style={{
-        fill: "#ecca86",
-      }}
-    />
-    <path
-      className="ak"
-      d="M79 202a2.38 2.38 0 0 0-1-2c-.5-.41-.44-1.15-.48-1.78-.06-.89-.06-1.79-.07-2.69a2.1 2.1 0 0 1 .74-1.78.81.81 0 0 0 .33-.75v-1.16c0-.78.2-1.62-.55-2.18a.62.62 0 0 1-.2-.57.17.17 0 0 1 .16-.06 4.61 4.61 0 0 1 .87 1.21v2.47a2.29 2.29 0 0 0 .1.94c.19.44-.25.73-.24 1.12s.26.53.17.85a16.39 16.39 0 0 0-.09 3.21c-.08.23-.06.55-.37.6a.57.57 0 0 1-.31-.55 16.71 16.71 0 0 1-.07-2A2.23 2.23 0 0 0 78 195a.6.6 0 0 1 .27-.7.64.64 0 0 0-.37.61 27.74 27.74 0 0 0 .17 4.37 2.3 2.3 0 0 0 .75 1c.14.13.32.21.38.43A1.82 1.82 0 0 1 79 202Z"
-    />
-    <path
-      d="M89.25 232.82c.31.16.59.33.58.81 0 .22 0 .49.41.31.19-.09.47.05.71.08 0 .2-.15.2-.27.29.65.44 1.19.93 1.8 1.34.24.16.49.35.79.18a.22.22 0 0 1 .16 0 .33.33 0 0 1 .16.28c.07 0-.19.17.06.2s.41.09.32.42a1 1 0 0 1-1.22 0 9.31 9.31 0 0 1-1.62-1.14c-.26-.68-.83-.86-1.38-1.1a.56.56 0 0 1-.26-.3 1.29 1.29 0 0 0 .54.71c.69.53 1.39 1 2.07 1.58a2.17 2.17 0 0 0 2 .45.33.33 0 0 1 .31.11 1.56 1.56 0 0 1-1.91.16c-.88-.57-1.74-1.2-2.59-1.85a2.12 2.12 0 0 1-.66-2.53Z"
-      style={{
-        fill: "#ebcb8a",
-      }}
-    />
-    <path
-      d="M116.88 223.38a2.06 2.06 0 0 1 .41 2.16c-.36 1.07-.79 2.12-1.21 3.17a1.65 1.65 0 0 1-2 1.15.42.42 0 0 1 .14-.71 1.13 1.13 0 0 0 .66-1.07 8.28 8.28 0 0 0 1.22-3.49c.16-.22.07-.73.54-.61a1.44 1.44 0 0 1 .14 1.39 7.9 7.9 0 0 1-.68 1.83c-.07.08-.17 0-.17 0-.13.6-.54 1.09-.49 1.75a.51.51 0 0 1-.25.44.89.89 0 0 0 .42-.53c.42-1.17.87-2.31 1.29-3.47a2.79 2.79 0 0 0-.17-1.72.41.41 0 0 1 .15-.29Z"
-      style={{
-        fill: "#eecb86",
-      }}
-    />
-    <path
-      d="M83.08 221.27a.84.84 0 0 0-.43-.16 2.05 2.05 0 0 1-1.54-1.6c-.25-1.07-.57-2.12-.81-3.19a2.08 2.08 0 0 1 .65-2c.53 0 .72.3.54 1a.85.85 0 0 0 .14.89 6.42 6.42 0 0 0 1 3.34c0 .2-.1.41.09.6a.36.36 0 0 1 0 .45c-.58 0-.77-.48-1-.9a7.32 7.32 0 0 1-.53-1.94c.17-.65-.26-1.18-.27-1.79 0-.2-.09 0-.13 0-.23-.14-.15-.34-.12-.53a7.44 7.44 0 0 0 .55 2.95c.09.28.14.58.21.88a1.52 1.52 0 0 0 .32.87c.31.56.87.68 1.35.93.05-.07.09.08-.02.2Z"
-      style={{
-        fill: "#ebc882",
-      }}
-    />
-    <path
-      className="af"
-      d="M95.28 201.07c.73.22 1.47.4 2.19.65a3.08 3.08 0 0 0 .89.19c.19 0 .43-.21.56.13a.26.26 0 0 0 .28.15 19.27 19.27 0 0 1 2.28 0 9.06 9.06 0 0 1 2.08-.08c.15 0 .31.09.45 0-.19 0-.38 0-.52-.16a.32.32 0 0 1 .11-.33.9.9 0 0 1 .51-.08c.26.22.63.2.83.53.12 0 .17 0 .19.13a1.75 1.75 0 0 1-1.71.54l-.16-.07a2.78 2.78 0 0 0-1.25 0c-.61 0-1.21.15-1.84.16.06.33.41.66 0 1a.33.33 0 0 1-.25-.07 5.14 5.14 0 0 0-2.2-1.1c-.25-.08-.53-.1-.7-.37s-.28-.3-.47-.37a5.39 5.39 0 0 1-1.46-.59c.04-.18.13-.21.19-.26Z"
-    />
-    <path
-      d="M120.48 210.33a12.53 12.53 0 0 1-.74 2.44c-.18-.2-.09-.42 0-.65.11-.42.22-.85.31-1.29-.15.66-.35 1.3-.48 2-.07.34.15 0 .17.08v.4c-.06.08 0 .22-.11.29s.1.33 0 .51a3.32 3.32 0 0 0 0 .63c0 .22 0 .44-.29.38a.86.86 0 0 0-.28 0c-.68.11-.75 0-.54-.74.56-1.56.85-3.22 1.3-4.82a3.36 3.36 0 0 1 .3-.82c.09-.16.19-.33.06-.52 0-.44.05-.45.57-.25a.45.45 0 0 1 .19.32c0 .15 0 .32.12.47a.42.42 0 0 1-.07.61 1.35 1.35 0 0 0-.51.96Z"
-      style={{
-        fill: "#e5a832",
-      }}
-    />
-    <path
-      className="ap"
-      d="M122 200.1c-.1.64-.16 1.3-.29 1.93s-.15 1.24.41 1.65a.21.21 0 0 1 0 .27 1.52 1.52 0 0 1-.84-.88c.15-.92.25-1.85.42-2.77-.19.89-.26 1.81-.43 2.71-.12.38.29.49.33.8a1.21 1.21 0 0 0 .09.53c.07.23 0 .4-.22.46s-.56.22-.86 0a1.23 1.23 0 0 1-.06-.63c.29-1.71.5-3.44.81-5.14a3.15 3.15 0 0 1 .23-.85.26.26 0 0 0-.09-.37.42.42 0 0 1 0-.38c.37-.21.6 0 .77.32a.39.39 0 0 1 0 .28 2.92 2.92 0 0 1-.21.28c.1-.06.12-.2.22-.25a.31.31 0 0 1 .17.55 1.72 1.72 0 0 0-.45 1.49Z"
-    />
-    <path
-      className="aq"
-      d="M78.6 204.54c.16-.39.6-.6.53-1.12.19 0 .2.17.26.31a.87.87 0 0 1 0 1.15.7.7 0 0 0-.08.43c0 .34.33.52.3.87a4.4 4.4 0 0 0 .5 3.15.46.46 0 0 1 0 .26.9.9 0 0 0-.06.43c0 .32-.13.36-.35.3a1.43 1.43 0 0 1-.64-1.06 9.05 9.05 0 0 1-.22-1.81c.17-.61-.09-1.22 0-1.67 0-.25-.11-.14 0-.17s0 0 0 .06a.33.33 0 0 1-.06.34c-.29 0-.22-.31-.24-.52a1.34 1.34 0 0 1 .17-.59c.06-.11.11-.23-.07-.27a.13.13 0 0 1-.04-.09Z"
-    />
-    <path
-      d="m93.53 236.13-.12-.25c.17-.38.48-.28.76-.29 1.52.46 3 1.11 4.47 1.72.28.12.45.48.79.45a.33.33 0 0 1 .08.27c0 .1-.08.18-.19.14s-.23 0-.26.06c-.29.53-.54.22-.78 0s-.45-.52-.79-.18c-.1-.39-.47-.32-.69-.52a.4.4 0 0 1 .22-.15 12.89 12.89 0 0 0-2-.82 8 8 0 0 0-1.49-.43Z"
-      style={{
-        fill: "#e4a62f",
-      }}
-    />
-    <path
-      className="as"
-      d="M121.38 204.7c.21 0 .3-.16.14-.37a.39.39 0 0 1 0-.54c.4.14.46.52.48.91a6.62 6.62 0 0 1-.18 2.08 2.62 2.62 0 0 0-.24 2c.08.24 0 .4-.17.57.14-.06.17-.22.29-.29a.21.21 0 0 1 .15 0l.06.07c-.22.79-.88.93-1.42 1.23-.14-.42-.15-.78.3-1s.43-.38.1-.69c-.06-.06 0-.25 0-.37a23.45 23.45 0 0 0 .34-2.78c0-.3-.05-.59.15-.82Z"
-    />
-    <path
-      className="at"
-      d="M116.63 224c-.16 0-.21.12-.21.28s0 .36-.26.37c-.14.26-.34.24-.55.16s-.17-.39-.08-.61c.48-1 .84-2.06 1.29-3.07a13.18 13.18 0 0 1 1.47-2.93.35.35 0 0 1 .11-.13.37.37 0 0 1 .28.17c.27.58.27.86-.17 1.2a1.37 1.37 0 0 0-.52 1l-.14.24c-.16-.12-.29-.07-.35.13-.26.82-.7 1.56-.9 2.41-.25.29.17.52.03.78Z"
-    />
-    <path
-      d="M108.38 237.21a1.28 1.28 0 0 1-1.6-.43.41.41 0 0 1 .4-.51c.57.31.86-.14 1.17-.54a9.56 9.56 0 0 0 2.24-1.84c.29-.26-.18-.42-.08-.66a.4.4 0 0 1 .52-.06c.06 0 .08.13.2 0 .45-.4.4 0 .43.34a4.27 4.27 0 0 1-1.8 2 6.73 6.73 0 0 0-1.82 1.39l-.1.05c-.09.08 0 .09 0 .1s.34 0 .44.16Z"
-      style={{
-        fill: "#eacb86",
-      }}
-    />
-    <path
-      className="av"
-      d="M98.81 202c-.1-.26-.35.11-.45-.13a2.46 2.46 0 0 1 1.87-.57h3.56a.94.94 0 0 1 .32.11.79.79 0 0 1-.51.15c-.35.21-.74.08-1.11.09a10.29 10.29 0 0 0-1.44.05 6.77 6.77 0 0 1 2.43.09c.27.08.58-.09.82.16-.12.14-.28.12-.43.12h-4.8a.33.33 0 0 1-.26-.07Z"
-    />
-    <path
-      className="aw"
-      d="M118 220.45c-.14-.44-.15-.75.33-1a.78.78 0 0 0 .27-1.19 15.17 15.17 0 0 0 .59-2.77c0-.2 0-.43.19-.57.13.07.22.07.18-.14s-.15-.49 0-.73a1 1 0 0 1 .47 1.11 13.2 13.2 0 0 1-.48 2.33c-.08.09-.19 0-.19 0 0 .57-.36 1.09-.25 1.68.06.38-.45.51-.32.93-.2.27-.55.16-.79.35Z"
-    />
-    <path
-      d="M115 228.08a.87.87 0 0 1-.58 1.14.46.46 0 0 0-.27.64h-.55c-.52.27-.75.86-1.13 1.3a20.14 20.14 0 0 0-1.34 1.68c-.2.15-.37.17-.44-.16a2.38 2.38 0 0 1 .73-1.11c.82-1 1.63-2 2.44-3a1.23 1.23 0 0 1 .51-.4c.18-.08.14-.22.1-.37a.28.28 0 0 1 .09-.25c.27.01.52.09.44.53Z"
-      style={{
-        fill: "#e6a831",
-      }}
-    />
-    <path
-      d="M87 229.68a.55.55 0 0 1 .22 0c.12 0 .35-.18.34.14s-.19.21-.36.19-.52 0-.79-.06c-.51.1-.75-.34-1-.71-.59-1.05-1.18-2.11-1.78-3.16V226a2.07 2.07 0 0 1 0-1c.25-.23.32 0 .4.18a1.67 1.67 0 0 0 .16.63 9.32 9.32 0 0 0 .82 1.4 12.15 12.15 0 0 1 .63 1.05 3.41 3.41 0 0 0 1.36 1.42Z"
-      style={{
-        fill: "#7d440d",
-      }}
-    />
-    <path
-      d="M80.13 209.6v-.25c0-.39.16-.53.5-.4a.91.91 0 0 1 .15.49c.38 1.71.7 3.43 1 5.14a2.42 2.42 0 0 1 0 .89.62.62 0 0 0 .11.58.2.2 0 0 1 0 .2.22.22 0 0 1-.33-.06c-.28-.23-.61-.36-.29-.91a.63.63 0 0 0-.37-1v-.14c.19-.19.13-.43.09-.65-.18-1-.43-2.06-.64-3.09.01-.3-.28-.46-.22-.8Z"
-      style={{
-        fill: "#e9ab31",
-      }}
-    />
-    <path
-      d="M78 189.13h-.14a3.55 3.55 0 0 1 0-1.3c.12-.42.17-.86.25-1.29.09-.58.16-1.18.3-1.75a1.91 1.91 0 0 1 1.36-1.5 1.4 1.4 0 0 1 1.62.61 2.19 2.19 0 0 0-1.39.1.18.18 0 0 1-.22 0c-.11-.13 0-.21.06-.33-.21 0-.22.38-.46.33s-.46.3-.52.64c-.25 1.41-.63 2.78-.79 4.2a.36.36 0 0 1-.07.29Z"
-      style={{
-        fill: "#e3bf59",
-      }}
-    />
-    <path
-      d="M122 200.1c-.22-.58-.25-1.08.33-1.45.2-.13.23-.4-.09-.5v-.28a25 25 0 0 1 .24-3.46c.34-.26.5-.1.52.29a8 8 0 0 1 0 1.47 9.14 9.14 0 0 0-.11 2.09c.17.4-.33.65-.21 1.05-.2.33-.62.39-.68.79Z"
-      style={{
-        fill: "#eccc86",
-      }}
-    />
-    <path
-      className="bc"
-      d="M80.13 209.6c.46.36.46.94.57 1.47.18.84.36 1.69.55 2.53.07.31 0 .5-.28.53-.18-.86-.37-1.73-.55-2.59 0-.18 0-.39-.19-.52s-.44-.06-.48-.38c.19-.1.21-.15.13-.36.27 0 .16-.21.1-.35s-.07-.31.15-.33Z"
-    />
-    <path
-      d="M120.43 184.34a.22.22 0 0 1 0-.11c.17 0 .31-.2.5-.1a1.06 1.06 0 0 1 .09.27 1.62 1.62 0 0 1 .63-.81c.22-.15.44-.32.71-.14 0 .05.07.13 0 .17-.56.65-.83 1.52-1.36 2.19-.17 0-.25.14-.35.24s-.23.25-.36 0a12.45 12.45 0 0 1 .14-1.71Z"
-      style={{
-        fill: "#f4d17d",
-      }}
-    />
-    <path
-      d="M108.38 237.21a1.83 1.83 0 0 1-.33 0c-.19 0-.51.08-.5-.2s.33-.13.51-.21l.31-.13c.65-.23 1.09-.84 1.67-1.21.11-.07.2-.19.32-.27a2.78 2.78 0 0 0 1.45-2.67.24.24 0 0 1 .11-.13 1.84 1.84 0 0 1-.27 2.07 15.08 15.08 0 0 1-3.27 2.75Z"
-      style={{
-        fill: "#c28727",
-      }}
-    />
-    <path
-      className="aw"
-      d="m119.62 213.48.1-.21a2.92 2.92 0 0 1 .57 3.13c-.24.81-.37 1.65-.58 2.47-.08.3-.14.67-.48.8a6 6 0 0 1 .34-1.63 7.27 7.27 0 0 0 .17-1 7.33 7.33 0 0 1 .2-.93 2.37 2.37 0 0 0-.3-2.26.23.23 0 0 1-.02-.37Z"
-    />
-    <path
-      d="m113.65 229.74.14.14-1.53 1.93c-.13.17-.32.3-.34.56v.06c-.19.16-.08.39-.11.58s.1.49-.22.54v-.59c-.28.69-.69.13-1 .32-.12 0-.19 0-.25-.15a.9.9 0 0 1 .18-.45.34.34 0 0 1 .12-.12.16.16 0 0 1 .2.07c0 .33.18.22.33.13.48-.53.87-1.16 1.37-1.67a6.75 6.75 0 0 1 1.11-1.35Z"
-      style={{
-        fill: "#cc8b25",
-      }}
-    />
-    <path
-      className="aq"
-      d="m79.75 210.66.48.36c-.93-.13-1.51-.87-1.66-2.14-.11-.92-.21-1.85-.29-2.78a2.09 2.09 0 0 1 .32-1.56l.06.05a20.42 20.42 0 0 0 .27 4.21 6.2 6.2 0 0 0 .14.75c.11.45.48.72.68 1.11Z"
-    />
-    <path
-      className="bg"
-      d="M93.53 236.13c.58-.25 1.08.07 1.6.28a6.88 6.88 0 0 0 1.6.7c.16.09.44 0 .43.38h-.36c-.54-.2-1.09-.38-1.62-.61a.59.59 0 0 0-.77.15h-.32c0-.14.25-.38-.11-.38-.12-.19-.3-.2-.47-.26s.03-.17.02-.26Z"
-    />
-    <path
-      d="M116.55 223.23c-.12-.24 0-.45.09-.66.27-.68.55-1.35.83-2 0-.11.09-.25.24-.22s.12.23.14.37a11 11 0 0 1-.7 1.66c-.21.31-.23.74-.47 1-.1.01-.16-.02-.13-.15Z"
-      style={{
-        fill: "#b77210",
-      }}
-    />
-    <path
-      d="M79.23 203.81c0-.14 0-.29-.1-.39L79 202v-1.39c.08-.15.17-.25.33-.11a4.16 4.16 0 0 1 .26 1.61c.07.41 0 .85.05 1.27s-.01.62-.41.43Z"
-      style={{
-        fill: "#bd7a16",
-      }}
-    />
-    <path
-      d="M107.24 236.4c-.29-.07-.32.24-.46.38a6.76 6.76 0 0 0-2.88 1.11c-.23-.52.14-.65.43-.76l1.51-.58a1.26 1.26 0 0 0 .54-.35 1.46 1.46 0 0 1 .72-.34c.22-.05.53-.07.58.2s-.29.23-.44.34Z"
-      style={{
-        fill: "#ce9226",
-      }}
-    />
-    <path
-      className="bk"
-      d="M83.7 226.09c.6.48.79 1.28 1.21 1.9a15.83 15.83 0 0 0 1 1.58 1.36 1.36 0 0 0 .57.39v.18a1.17 1.17 0 0 1-1-.73c-.58-1-1.12-2-1.67-2.95a.47.47 0 0 1-.11-.37Z"
-    />
-    <path
-      className="as"
-      d="M122.08 203.94v-.26a1.41 1.41 0 0 1 .43 1.3c-.12 1.25-.29 2.49-.45 3.72a1.55 1.55 0 0 1-.17.4h-.05a2.42 2.42 0 0 1 .05-1.36c0-.47.08-.93.09-1.4.08-.83.02-1.62.1-2.4Z"
-    />
-    <path
-      d="M106.44 236.28c0 .29-.23.3-.4.36l-1.64.62c-.26.11-.56.19-.5.63v.24l-.2.06c0-.12-.07-.26-.24-.15a.56.56 0 0 0-.38 0c-.28 0-.29-.12-.2-.35.37-.5 1-.5 1.43-.73s1.12-.41 1.67-.63c.14-.12.3-.22.46-.05Z"
-      style={{
-        fill: "#a55f06",
-      }}
-    />
-    <path
-      d="M83.08 221.27A.25.25 0 0 0 83 221c0-.23.09-.31.28-.28a1.19 1.19 0 0 1 .41.59c.28.64.51 1.31.74 2 .1.28.23.66-.28.65-.09 0-.12-.14-.16-.23v-.21c-.14-.61-.46-1.13-.65-1.71a.87.87 0 0 1-.26-.54Z"
-      style={{
-        fill: "#bc7610",
-      }}
-    />
-    <path
-      d="M122.4 183.5c-.28-.09-.49.12-.72.22l-.3-.19c-.1-.08-.29-.09-.24-.33.65-.35 1.31-.61 1.93.07-.11.1-.18.27-.35.25s-.21 0-.32-.02Z"
-      style={{
-        fill: "#ecc461",
-      }}
-    />
-    <path
-      d="m95.28 201.07-.11.13a.22.22 0 0 1-.34 0c-.22-.24-.6-.11-.79-.42.09-.12 0-.26 0-.38a2.1 2.1 0 0 0-.81-.86c-.16-.13-.39-.18-.44-.45l.13-.13a3 3 0 0 1 1.83 1.81.46.46 0 0 0 .53.3Z"
-      style={{
-        fill: "#ebc269",
-      }}
-    />
-    <path
-      d="M119.72 212.87c-.12.07-.11.29-.35.32a14.84 14.84 0 0 1 .77-2.93c.13.89-.37 1.66-.4 2.51a.3.3 0 0 1-.02.1Z"
-      style={{
-        fill: "#a86102",
-      }}
-    />
-    <path
-      d="M119.89 217a8.78 8.78 0 0 1-.66 2.68l-.44.41c-.32-.11-.26-.29-.08-.5s.32-.16.33-.4 0-.15 0-.21c.28-.54.31-1.17.55-1.72.1-.14.1-.32.3-.26Z"
-      style={{
-        fill: "#bd8726",
-      }}
-    />
-    <path
-      className="br"
-      d="M122.63 199.31c-.31-.05-.23-.24-.11-.44s.26-.24.25-.46-.09-.3.07-.4.06 0 .08.05c-.07-.3-.2-.61.18-.79v1.42a1.37 1.37 0 0 1-.47.62Z"
-    />
-    <path
-      className="bk"
-      d="m84 223.7.13.13c.52.43.55.61.25 1.39-.1.13-.2.08-.3 0a2.69 2.69 0 0 0-.33-.6c.01-.33.25-.57.25-.92Z"
-    />
-    <path
-      className="bs"
-      d="M121.14 183.2c.06.18.25.12.35.23a1.45 1.45 0 0 1-.61.68c-.09.27-.23.37-.43.12a1.51 1.51 0 0 1 .69-1.03Z"
-    />
-    <path
-      d="M86.5 194a.61.61 0 0 0 .15-.44c0-.24-.08-.5.13-.69a2.73 2.73 0 0 1 .07 1.42c-.47.41-.42-.16-.57-.37.13-.1.16.08.22.08Z"
-      style={{
-        fill: "#e7ba4d",
-      }}
-    />
-    <path
-      className="bs"
-      d="M120.43 184.34a6.53 6.53 0 0 0 .09 1.82c0 .46-.13.87-.59 1h-.11a4 4 0 0 0 .5-2.3.68.68 0 0 1 .11-.52Z"
-    />
-    <path
-      className="av"
-      d="m105.1 202.17-.16-.16a4.38 4.38 0 0 1 1.42-.46.2.2 0 0 1 .25 0c-.07.42-.42.35-.65.46a2.5 2.5 0 0 1-.7.23.23.23 0 0 1-.16-.07Z"
-    />
-    <path
-      className="bu"
-      d="M83.3 221.78a3.64 3.64 0 0 1 .65 1.71 5.71 5.71 0 0 1-.65-1.71Z"
-    />
-    <path
-      className="bv"
-      d="M106.59 201.56h-.23a3.33 3.33 0 0 1 1-.35l.11.13c-.18.46-.47.46-.88.22Z"
-    />
-    <path
-      className="at"
-      d="M116.65 223.36c.17-.33.2-.76.5-1 0 .38-.33.64-.27 1v.27c-.18-.01-.28-.07-.23-.27Z"
-    />
-    <path
-      className="bk"
-      d="M83.71 225v1a1.3 1.3 0 0 1 0-1.33.24.24 0 0 1 0 .33Z"
-    />
-    <path
-      className="ak"
-      d="M78.09 186.55a2.19 2.19 0 0 1-.25 1.29 2.71 2.71 0 0 1 .25-1.29Z"
-    />
-    <path
-      d="M83.71 225v-.31a.17.17 0 0 1 0-.1c.38 0 .22.36.34.53s0 .26-.13.26c-.04-.13.08-.38-.21-.38Z"
-      style={{
-        fill: "#efbf59",
-      }}
-    />
-    <path className="ak" d="M77.84 187.84v1.3a1.63 1.63 0 0 1 0-1.3Z" />
-    <path
-      className="ab"
-      d="M83.25 190.9a.7.7 0 0 1-1 .11.23.23 0 0 1-.15-.28 2.29 2.29 0 0 1 .08-.28c-.06 0-.07.09-.1.15s-.1.36-.31.27-.21-.27-.21-.47c0-.36.12-.77-.23-1a.43.43 0 0 1 0-.35c.26-.09.34-.25.18-.53l-.26-.4c-.14-.2 0-.58-.31-.68a3.89 3.89 0 0 0-.25-.5 1.73 1.73 0 0 1-.14-.31c-.2-.67-.27-1.43-1-1.68-.31-.48-.27-.63.24-.88H80c0 .2.13.24.2.37a11 11 0 0 1 1.09 2.23c.44 1 .83 2 1.24 3a2.1 2.1 0 0 1 .2 1c.14-.19 0-.46.23-.59.21.18-.09.65.29.82Z"
-    />
-    <path
-      className="bx"
-      d="M97.07 202.15a6.28 6.28 0 0 0 1.45.52 3.33 3.33 0 0 1 1.52.94.14.14 0 0 1 0 .18.36.36 0 0 1-.52.05c-1.09-.85-2.42-1-3.63-1.44a25.89 25.89 0 0 1-2.78-1c0-.26.14-.38.37-.33l1.48.41a4.26 4.26 0 0 1 2.11.67Z"
-    />
-    <path
-      d="M83 190v.7c0 .23-.06.26-.2.24s-.21-.08-.16-.26a1.87 1.87 0 0 0-.26-1.27c-.71-1.63-1.34-3.31-2.14-4.9v-.11c.32 0 .5.2.63.52s.33.71.49 1.06a4 4 0 0 1 .86 1.95Z"
-      style={{
-        fill: "#9c610f",
-      }}
-    />
-    <path
-      d="M85.26 193.94c.69 0 .88.71 1.33 1.06a4.62 4.62 0 0 1 1.56 1.82l1.15 1.47a1.43 1.43 0 0 1 .34.71c0 .16 0 .33-.12.38s-.28-.27-.43-.39c-1.08-1.6-2.41-2.94-3.62-4.4-.1-.12-.24-.21-.28-.39a.27.27 0 0 1 .07-.26Z"
-      style={{
-        fill: "#9b5f0e",
-      }}
-    />
-    <path
-      d="m89.18 198.87.15.17c.07.08 0 .29.15.26s.13-.18.16-.3l.33.42-.25.25a1.24 1.24 0 0 1-.71.1c-.41-.16-.45-.76-.85-.94a.58.58 0 0 1-.1-.12c-.09-.19-.31-.1-.43-.24a16.6 16.6 0 0 0-2.14-2.63 1.38 1.38 0 0 0-.67-.09c-.2 0-.42 0-.53-.26.15-.28.4-.18.61-.21s.46.06.65-.14a.49.49 0 0 1 .15-.1c.23 0 .34.17.4.4a.78.78 0 0 0 .31.58c.57.57 1 1.37 1.68 1.75.41.31.63.85 1.09 1.1Z"
-      style={{
-        fill: "#cc8e24",
-      }}
-    />
-    <path
-      d="M94.14 200.43c0 .13.22.32-.1.33l-4.62-3 .08-.16c.77 0 1.26.73 1.94.94.33 0 .56.22.82.38.48.31 1 .64 1.44 1 .18.08.43.19.44.51Z"
-      style={{
-        fill: "#9d661c",
-      }}
-    />
-    <path
-      d="M84.75 191.45a14.76 14.76 0 0 0-1.47-1.94.25.25 0 0 1 0-.34c.09-.12.21 0 .3 0a3.24 3.24 0 0 1 .85 1c.21.28.4.57.59.85a.26.26 0 0 1 .15.36c-.17.21-.3.09-.42.07Z"
-      style={{
-        fill: "#a36418",
-      }}
-    />
-    <path
-      d="M94.14 200.43c-.84-.74-1.85-1.15-2.7-1.87.25-.29.44-.05.64.08s.5.28.74.43a5.07 5.07 0 0 1 1.26 1c.12.08.25.19.06.36Z"
-      style={{
-        fill: "#fbeab9",
-      }}
-    />
-    <path
-      d="M84.75 191.45c.12-.19.22 0 .33 0a.15.15 0 0 1 .21 0 4.93 4.93 0 0 1 1.1 1.83l.11.72-.22-.08a24.33 24.33 0 0 0-1.53-2.47Z"
-      style={{
-        fill: "#7a4a0a",
-      }}
-    />
-    <path
-      d="M85.26 193.94v.22c-.12.22.06.28.15.38s.35.25.29.53a.89.89 0 0 1-1.12.08c-.08-.13.24-.29 0-.46s.14-.36.06-.58a.76.76 0 0 1 .24-.63c.21-.05.21.11.22.27Z"
-      style={{
-        fill: "#cc9022",
-      }}
-    />
-    <path
-      d="M86.39 193.32c-.48-.58-.79-1.31-1.31-1.85l-.11-.39c0-.16.11-.33.25-.14A10.31 10.31 0 0 1 86.5 193a.25.25 0 0 1-.11.32Z"
-      style={{
-        fill: "#fdf1c8",
-      }}
-    />
-    <path
-      d="M88.15 196.82 86.59 195c0-.11 0-.28.21-.14a7.06 7.06 0 0 1 1.48 1.73c.11.19-.02.2-.13.23Z"
-      style={{
-        fill: "#fcecbe",
-      }}
-    />
-    <path
-      d="M97.07 202.15a13.31 13.31 0 0 1-2.13-.71l-.11-.26h.34l1.86.64c.19.08.45.18.04.33Z"
-      style={{
-        fill: "#f8e9bb",
-      }}
-    />
-    <path
-      d="M82.17 187.94a15.39 15.39 0 0 1-.86-1.95c0-.16.08-.34.24-.09a10.13 10.13 0 0 1 .83 1.87c.1.23-.07.23-.21.17Z"
-      style={{
-        fill: "#fbeec0",
-      }}
-    />
-    <path
-      d="m93.46 201-.36.25c-.27.16-.4 0-.44-.26s.39-.25.46-.51c.12.19.47.17.34.52Z"
-      style={{
-        fill: "#eeddb5",
-      }}
-    />
-    <path
-      d="m85 193.79-.22-.27c.16-.31.62.1.7-.37.24.64.24.64-.48.64Z"
-      style={{
-        fill: "#f2e5ba",
-      }}
-    />
-    <path
-      d="M80.75 184c.53.16.58.51.6.89Z"
-      style={{
-        fill: "#9d6725",
-      }}
-    />
-    <path
-      d="M113.55 198.36c-.2-.53.34-.66.44-1s.28-.48.47-.67c.5-.51 1-1 1.48-1.55s.83-.94 1.29-1.35a1.69 1.69 0 0 1 1.07-.55.43.43 0 0 1 .36.43c0 .25-.23.23-.41.25a1.22 1.22 0 0 0-.66.47l-3.29 3.55a2.72 2.72 0 0 1-.52.5.2.2 0 0 1-.23-.08Z"
-      style={{
-        fill: "#9c5f0d",
-      }}
-    />
-    <path
-      className="bs"
-      d="M118.3 193.29a4.57 4.57 0 0 0-1.83 1.47c-.15.16-.32.3-.48.45-.1-.05-.06-.18-.11-.26.33-.34.64-.71 1-1a1.13 1.13 0 0 0 .41-1.18c-.12-.22 0-.41.08-.59.37-.93.8-1.82 1.21-2.73.08-.18.14-.41.4-.36 0 .08.15.13.17 0 .11-.44.48-.74.55-1.2.12-.05.2-.22.36-.11.12.31-.08.51-.2.74a1 1 0 0 0-.12.94c-.09.29 0 .45.27.41a.28.28 0 0 1 .12.12c-.25.44-.56.23-.84.08s-.41-.14-.51.14c-.34.86-.77 1.67-1.13 2.53.31.06.62-.14.82.21a.31.31 0 0 1-.17.34Z"
-    />
-    <path
-      d="M119 189.08a7.15 7.15 0 0 0-.77 1.54c-.32.7-.61 1.42-.92 2.14-.28.3-.21.86-.68 1.12-.07-1 .6-1.58.73-2.38a5 5 0 0 1 .94-2.19c.33-.16.4-.7.85-.72.04.21-.1.34-.15.49Z"
-      style={{
-        fill: "#9e651d",
-      }}
-    />
-    <path
-      d="M108.89 200.83c-.17-.37.13-.52.3-.69s.63-.39.85-.71a1.67 1.67 0 0 1 1.13-.8c.23.3-.17.43-.14.67-.35.48-.9.66-1.28 1.1-.25.14-.44.4-.72.49a.14.14 0 0 1-.14-.06Z"
-      style={{
-        fill: "#aa7221",
-      }}
-    />
-    <path
-      d="M111.22 198.7a5 5 0 0 0-1.18.73c-.07 0-.19-.08-.14-.13a9.48 9.48 0 0 1 2.65-2 .23.23 0 0 1 .3.11 4 4 0 0 1-1.63 1.29Z"
-      style={{
-        fill: "#f9e8bc",
-      }}
-    />
-    <path
-      d="M111.22 198.7a15.09 15.09 0 0 1 1.63-1.26l.64-.13-1.13 1v.08c-.46.23-.77.78-1.32.84-.19-.3.16-.33.18-.53Z"
-      style={{
-        fill: "#8e5c1b",
-      }}
-    />
-    <path
-      className="bv"
-      d="M108.89 200.83a.14.14 0 0 1 .09 0c.45-.28.93-.14 1.4-.2a.31.31 0 0 1 .35.23c-.1.3-.34.29-.56.31a1.44 1.44 0 0 0-.65.07c-.59.15-1.14.54-1.78.39a.26.26 0 0 1 0-.29c.37-.2.82-.18 1.15-.51Z"
-    />
-    <path
-      d="M115.88 195c.1.06.1.16.11.26a26.78 26.78 0 0 1-2 2.12c-.14-.12 0-.22 0-.32a12.7 12.7 0 0 1 1.89-2.06Z"
-      style={{
-        fill: "#faecbb",
-      }}
-    />
-    <path
-      d="M118.27 189.31a18.5 18.5 0 0 1-.94 2.19c-.15.06-.29.05-.22-.18a7 7 0 0 1 1-2.07c.04-.12.14-.18.16.06Z"
-      style={{
-        fill: "#fcf1c8",
-      }}
-    />
-    <path
-      d="M107.73 201.33v.23a4.3 4.3 0 0 1-2.3.88c-.12 0-.25-.08-.23-.27l1.39-.61a1.34 1.34 0 0 0 .93-.23Z"
-      style={{
-        fill: "#f8ebbe",
-      }}
-    />
-    <path
-      d="m120 187.82-.33.11a1.66 1.66 0 0 0 .24-.75 4.93 4.93 0 0 0 .59-1c.1-.16.19-.32.29-.47s.15-.07.17.07a3 3 0 0 1-.96 2.04Z"
-      style={{
-        fill: "#fbf0bc",
-      }}
-    />
-    <path
-      d="M113.05 198.6c-.25-.45-.53 0-.8 0-.06-.14.08-.18.11-.27a1.08 1.08 0 0 0 .64-.2c.1-.07.18-.27.32-.18s.08.26.08.4-.16.26-.35.25Z"
-      style={{
-        fill: "#f8e3b9",
-      }}
-    />
-    <path
-      className="cx"
-      d="M113.55 198.36h.19c.26.31.31.55-.15.65h-.09c-.09 0-.19 0-.24-.13a.74.74 0 0 0 .29-.52Z"
-    />
-    <path
-      className="cy"
-      d="M121.38 193.76a1.93 1.93 0 0 0-.45 1.3v.75a.11.11 0 0 1-.17 0 4.52 4.52 0 0 1 0-1.6 1.51 1.51 0 0 1 .15-.59 1.38 1.38 0 0 0-.16.69.45.45 0 0 1-.08.26.18.18 0 0 1-.16 0 .55.55 0 0 1-.08-.52c.09-.25-.13-.11-.16-.19-.12-.3 0-.53.15-.77a5.6 5.6 0 0 0 .5-1.27c.09-.35.25-.67.37-1a2.63 2.63 0 0 0-.06-1.57.88.88 0 0 1 0-.28c0-.41-.4-.7-.2-1.15.11-.27 0-.64.26-.85.12 0 .15.1.15.21s0 .29.18.22c0 .47.17.71.63.52a.37.37 0 0 1 .47.15c-.1.14-.25.1-.38.13a6.79 6.79 0 0 0-.39 2.1c-.11 1.08-.25 2.16-.37 3.24a.42.42 0 0 1-.2.22Z"
-    />
-    <path
-      className="cz"
-      d="M120 187.82a12.47 12.47 0 0 0 1-2.06c.24-.14.3-.44.39-.68a6.52 6.52 0 0 1 1-1.58c.12-.18.22 0 .32 0 .19.33 0 .49-.22.62v.1a45.28 45.28 0 0 0-2.25 4.52 3.51 3.51 0 0 1-.44.75.18.18 0 0 1-.16 0c-.22-.24-.15-.48 0-.75a4.07 4.07 0 0 0 .36-.92Z"
-    />
-    <path
-      d="M121.38 193.76a15.25 15.25 0 0 1 .38-3.16 3.17 3.17 0 0 0 0-1 1.57 1.57 0 0 1 .5-1.43c0 .11.23.2.16.31-.62 1-.35 2.13-.55 3.19-.15.78-.07 1.6-.28 2.38Z"
-      style={{
-        fill: "#a0650f",
-      }}
-    />
-    <path
-      className="db"
-      d="m122.76 188 .06-2.47c0-.2.07-.43-.1-.59s0-.34 0-.51a.92.92 0 0 1 .22-.26c.15-.08.31-.21.35.11v.1a9.91 9.91 0 0 0 0 1.54 1.52 1.52 0 0 1 0 .21 8.17 8.17 0 0 1 0 1.4c0 .34-.09.69-.47.78a.34.34 0 0 1-.06-.31Z"
-    />
-    <path
-      d="M122.91 188.3c.31-.24.29-.62.31-1s0-.9.07-1.34a5.4 5.4 0 0 1 .1-1.7c.15-.07.14.05.15.16a29.89 29.89 0 0 1 0 3.14c0 .93-.45 1.42-1.18 1.52-.09-.48.58-.33.55-.78Z"
-      style={{
-        fill: "#743c09",
-      }}
-    />
-    <path
-      d="M123.14 197.26c-.2.32 0 .7-.16 1 0 0-.06 0-.07-.07a1.13 1.13 0 0 1 0-.18 4.59 4.59 0 0 1-.06-1.48 9.2 9.2 0 0 0 .08-2 2.29 2.29 0 0 0-.62-1.24c-.13-.22-.38-.41-.15-.72a2 2 0 0 1 1.17 2 17.26 17.26 0 0 0-.09 1.9 1.61 1.61 0 0 1-.07.69Z"
-      style={{
-        fill: "#743c07",
-      }}
-    />
-    <path
-      d="M122.59 189.8c-.19.79-.06 1.61-.25 2.4 0 .1 0 .24-.14.23s-.17-.13-.16-.27c.08-.8.15-1.6.23-2.4 0-.14 0-.34.18-.31s.13.22.14.35Z"
-      style={{
-        fill: "#b56f0f",
-      }}
-    />
-    <path
-      className="cy"
-      d="M122.31 193.25c.53.24.6.78.62 1.3s0 1.28 0 1.93c-.18 0-.14-.11-.09-.23v-1.54c0-.34-.13-.33-.34-.28a.37.37 0 0 0-.45-.26V194a7 7 0 0 1 .26-.75Z"
-    />
-    <path
-      className="cz"
-      d="m123 184.16-.09.26a4.64 4.64 0 0 0-.59 1c-.1.16-.2.47-.44.14a2.94 2.94 0 0 1 .66-1.4Z"
-    />
-    <path
-      d="M122.31 193.25c.09.33.19.66-.24.78a1.58 1.58 0 0 1 .06-1.39c-.13.29.21.36.18.61Z"
-      style={{
-        fill: "#ca891b",
-      }}
-    />
-    <path
-      className="br"
-      d="M123.14 197.27a14.9 14.9 0 0 1 .16-2.59 10.72 10.72 0 0 1-.16 2.59Z"
-    />
-    <path
-      className="db"
-      d="M123.39 184.29a6.06 6.06 0 0 1-.1 1.7l-.05-.07a7.88 7.88 0 0 1 0-1.61Z"
-    />
-    <path
-      className="dg"
-      d="M114.51 227.55v.26c-.68.52-1.1 1.37-1.82 1.84a3.44 3.44 0 0 0-1.49 1.88c0 .48-.49.69-.64 1.1h-.06c-.15-.19-.15-.19-.51.12-.73.57-1.42 1.2-2.12 1.85.14.22.59.21.41.67-.07.08-.16.16-.25.14-.73-.09-1.35.44-2 .45-.52-.2-1 .14-1.43.26-.18.05-.35.16-.53.22s-.15 0-.22.07.14 0 .15.12c-.33.47-.86.44-1.3.64a.37.37 0 0 1-.18 0 4.9 4.9 0 0 0-2.6-.26c-.23 0-.33.1-.24.4s.19.62-.19.78v-.25c-.74-.67-1.72-.81-2.5-1.37a2 2 0 0 0-1.84-.68c-.4 0-.75-.22-1.12-.28-.19-.12-.08-.45-.3-.54-.6-.56-1.3-1-1.91-1.51-.34-.31-.35.3-.59.21s-.23-.14-.32-.26a10.07 10.07 0 0 1-1.07-1.76 1.23 1.23 0 0 0-.26-.67.7.7 0 0 1-.17-.61.15.15 0 0 1 .13-.1.46.46 0 0 1 .23.18 2.28 2.28 0 0 0 1.59 1.11 2.47 2.47 0 0 1 1.07.59c.53.42 1.06.83 1.61 1.21a2.65 2.65 0 0 1 1.14 1.3 1 1 0 0 0 .63.6c.58.22 1.16.41 1.74.62a.91.91 0 0 0 .83-.12 2.18 2.18 0 0 1 1.27-.28 11.82 11.82 0 0 0 1.88-.06 2.61 2.61 0 0 1 1.78.33.81.81 0 0 0 .71.08c.63-.2 1.23-.48 1.84-.72a.81.81 0 0 0 .44-.56 2.67 2.67 0 0 1 .82-1.13c.67-.57 1.36-1.11 2-1.71a1.84 1.84 0 0 1 1-.55 1.94 1.94 0 0 0 1.35-1c.19-.25.36-.5.57-.73a2.76 2.76 0 0 0 .55-1.54 13.2 13.2 0 0 1 1.72-4.64.62.62 0 0 1 .3-.28c.79-.26 1.06-1 1.31-1.74a8.53 8.53 0 0 1 .4-1 .78.78 0 0 0 0-.68 3.66 3.66 0 0 1-.14-2 26.09 26.09 0 0 1 .59-2.62c0-.14.08-.29.24-.33a.38.38 0 0 1 .21.29 4.89 4.89 0 0 1-.32 1.65c-.08.25.13.06.14.16.16.3 0 .57 0 .86a3.68 3.68 0 0 0-.15 1.36 7.22 7.22 0 0 1 .26-1.6c.06-.21.07-.48.34-.52.29.36 0 .75 0 1.1.3.16.74-.64.89.21v.06a4.19 4.19 0 0 0-1.18 2c-.1.26-.13.6-.46.66a.13.13 0 0 1 0-.06 3.62 3.62 0 0 0-.63 1.91c0 .57-.42 1-.39 1.6 0 .17-.15.33 0 .51-.67-.17-.79.4-1 .9-.12.29-.09.76-.55.75a.32.32 0 0 1-.06-.3 12.4 12.4 0 0 1 1-2.44c-.26.15-.24.67-.58.86-.2-.13-.2.09-.25.2a3.09 3.09 0 0 0-.42 1.58c0 .6-.37 1.07-.41 1.65a7.44 7.44 0 0 1 .4-1.15c.08-.17.11-.43.37-.39.23.31 0 .59 0 .91-.06.07.27-.21.29.19Z"
-    />
-    <path
-      className="dh"
-      d="M89.42 230.26v.13c-.43 0-.59-.44-.82-.73s-.15-.06-.23 0c-.24 0-.34-.24-.46-.43a1.78 1.78 0 0 0-.74-.65l-.09-.15c.26-.16.43-.35.23-.7-.39-.8-.89-1.52-1.3-2.32-.18 0-.17.36-.38.34a.39.39 0 0 1-.28-.19c-.12-1-.57-1.81-.75-2.75.24-.18.11-.39 0-.59a5 5 0 0 1-.43-1.3.25.25 0 0 1 .2-.16c.43.3.47.85.65 1.31.27.66.48 1.32 1.2 1.58.39.14.63.59.86 1 .48.82.92 1.66 1.4 2.49a2 2 0 0 1 .29 1.15 1.82 1.82 0 0 0 .48 1.56.41.41 0 0 1 .17.41Z"
-    />
-    <path
-      d="M84.81 225.81a.15.15 0 0 0 0 .21 24.72 24.72 0 0 0 1.5 2.6c.22.3.34.18.44-.1a1.37 1.37 0 0 0 .33 1.16 1.52 1.52 0 0 1-1.22-.89c-.3-.48-.59-1-.89-1.46a2.87 2.87 0 0 0-.81-1.39.45.45 0 0 1-.18-.54c0-.09.08-.17.13-.26l.2.12c.12.2.56.06.5.55Z"
-      style={{
-        fill: "#ebcc8b",
-      }}
-    />
-    <path
-      d="M87.17 228.55c.37-.1.55.16.75.43s.26.47.49.61a5.49 5.49 0 0 1 1.42 2c.3.63.91 1 1 1.8 0 .18 0 .4-.27.47a1.79 1.79 0 0 0-.47-1.26c-.96-1.35-1.94-2.7-2.92-4.05Z"
-      style={{
-        fill: "#9a5f0d",
-      }}
-    />
-    <path
-      d="M86.83 228.24c.08-.35.55-.21.59-.6.27.76.27.76-.34.76Z"
-      style={{
-        fill: "#f7e5b9",
-      }}
-    />
-    <path
-      d="M90.11 233.08c.19.21.32.34.15.62-.16-.14-.26-.29-.15-.62Z"
-      style={{
-        fill: "#a66404",
-      }}
-    />
-    <path
-      d="M103.44 238c.15 0 .06-.36.23-.26s0 .29 0 .45a1.94 1.94 0 0 1-2 1.12c-.89-.07-1.78 0-2.67 0-.57 0-.74-.14-1.06-.79a1.93 1.93 0 0 0 1.93.45.25.25 0 0 1 .33 0c.69 0 1.37-.06 2.05-.11a1.39 1.39 0 0 0 1-.6c.13-.07.09-.21.19-.26Z"
-      style={{
-        fill: "#7c450a",
-      }}
-    />
-    <path
-      d="M103.44 238c0 .1 0 .21-.14.24-.28-.19-.64 0-.9-.25l-.24-.26c0-.31.39-.32.38-.64h.1a.33.33 0 0 1 .17.25v.4c0 .13.14.2.27.13s.34-.14.36.13Z"
-      style={{
-        fill: "#ca902e",
-      }}
-    />
-    <path className="do" d="M100.21 239h-.33l-.9-.09c.42-.06.84-.23 1.23.09Z" />
-    <path
-      d="M78.5 195v-.07c.47-1.54 0-3.13.21-4.69.11 0 .16 0 .17.15a19 19 0 0 1 0 2.13c0-.59.06-1.3 0-2 .1-.43-.36-.58-.38-.95s-.18-.8.17-1.11c.66-.23.8-.12.78.65-.06 1.59-.07 3.18-.11 4.77a2.65 2.65 0 0 1-.15 1c-.07.11-.17.34-.28.14s-.27-.02-.41-.02Z"
-      style={{
-        fill: "#e9ab39",
-      }}
-    />
-    <path
-      d="m78.84 190.3-.13-.11a2.82 2.82 0 0 1-.76-1.06 22.51 22.51 0 0 1 .69-4.1 1.59 1.59 0 0 1 .51-1.16c.17-.12.24-.1.22.14s-.14.35-.22.52a8.9 8.9 0 0 0-.37 1.93 8.44 8.44 0 0 0-.2 3.09c0 .32.67.25.35.76-.03.06-.06.06-.09-.01Z"
-      style={{
-        fill: "#7a4308",
-      }}
-    />
-    <path
-      d="m79.29 200.47-.25.15a1.9 1.9 0 0 1-1.1-1.18 25.44 25.44 0 0 1-.14-3.6c0-.2 0-.44.18-.56a2.16 2.16 0 0 1 .14 1.58 11.22 11.22 0 0 0 .26 2.59c.28.19.47.54.81.65.12.09.22.19.1.37Z"
-      style={{
-        fill: "#7d480e",
-      }}
-    />
-    <path
-      d="M78 195.3a15.36 15.36 0 0 0-.11 3.32v.82a1.58 1.58 0 0 1-.18-.79c0-1.14 0-2.27-.06-3.4a1.38 1.38 0 0 1 .27-.95c.08-.1.16-.31.31-.2s0 .27 0 .38a.65.65 0 0 0-.18.57.19.19 0 0 1-.05.25Z"
-      style={{
-        fill: "#ad7316",
-      }}
-    />
-    <path
-      d="M78 195.3v-.25a1.65 1.65 0 0 1 .07 1.81 7.4 7.4 0 0 1-.07-1.56Z"
-      style={{
-        fill: "#f7e4bc",
-      }}
-    />
-    <path
-      d="M79.06 200.18a1.06 1.06 0 0 1-.68-.73c.21-.1.07-.42.26-.53.07.14.16.26.33.24a1.34 1.34 0 0 1 .09 1.02Z"
-      style={{
-        fill: "#ebbd56",
-      }}
-    />
-    <path
-      d="M78.5 195c.15-.13.18-.38.39-.45.16.15 0 .43.16.55a1 1 0 0 1 0 .65c-.11 0-.2 0-.26-.1s-.42-.35-.29-.65Z"
-      style={{
-        fill: "#bf811c",
-      }}
-    />
-    <path
-      d="M94 236.65c.2.12.4-.45.54-.07s-.32.26-.43.45a1.47 1.47 0 0 1-1.63-.14c-.89-.64-1.76-1.32-2.64-2a.84.84 0 0 1-.18-.15 1.17 1.17 0 0 1-.35-.9c0-.18.18-.07.24 0s0 .39.23.48a4.68 4.68 0 0 1 1.38 1.06c.45.31.89.63 1.34.93s.92.69 1.5.34Z"
-      style={{
-        fill: "#875213",
-      }}
-    />
-    <path
-      d="M91.14 235.45a4.85 4.85 0 0 1-1.38-1.06 1.63 1.63 0 0 1 1.38 1.06Z"
-      style={{
-        fill: "#f5e2bb",
-      }}
-    />
-    <path
-      className="bg"
-      d="M93.32 235.85v-.47c.17-.12.32 0 .48 0s.42 0 .42.34a1.3 1.3 0 0 0-.82.13Z"
-    />
-    <path
-      d="m116.65 223.36.19.29a2.12 2.12 0 0 1 .12 2.1c-.4 1-.77 2-1.16 3.05a1.05 1.05 0 0 1-1 .81c.06-.42.59-.32.56-.74a4.68 4.68 0 0 1 .61-1.65c.2-.58.41-1.15.6-1.72a1.9 1.9 0 0 0 0-1.48c-.33-.57-.33-.57-.08-.79Z"
-      style={{
-        fill: "#864e0d",
-      }}
-    />
-    <path
-      d="M116 227.22a6.07 6.07 0 0 1-.61 1.65 3.42 3.42 0 0 1 .53-1.81Z"
-      style={{
-        fill: "#f6e5c2",
-      }}
-    />
-    <path
-      className="bu"
-      d="M82.55 219.43c.13-.5.33-.48.58-.1a.55.55 0 0 1 .08.15c.69 1.6 1.21 3.29 1.86 4.91a1.61 1.61 0 0 1 .13.68c0 .21 0 .38.21.47l.21.12a.76.76 0 0 1-.81.15c0-.2-.11-.29-.3-.25s-.21-.14-.25-.3c.13-.5.28-1-.16-1.43.32 0 .31-.13.22-.37a18.58 18.58 0 0 0-1.11-2.73c-.2-.06-.43 0-.56-.27.04-.35-.4-.64-.1-1.03Z"
-    />
-    <path
-      d="M82.65 220.46c.2.06.43 0 .56.27L83 221a1.71 1.71 0 0 1-1.36-.9c-.17-.28-.16-.64-.34-.91a2.74 2.74 0 0 1-.2-.9c-.14-.52-.23-1.05-.39-1.57-.07-.22-.2-.48.06-.67a2.81 2.81 0 0 1 .49 1.58c.17.59.34 1.2.52 1.79a1.41 1.41 0 0 0 .87 1.04Z"
-      style={{
-        fill: "#7e460c",
-      }}
-    />
-    <path
-      className="eb"
-      d="M80.78 216.06c-.23.29 0 .56 0 .83a6.44 6.44 0 0 1 .26 1.39 9.44 9.44 0 0 1-.61-2.23 1.6 1.6 0 0 1 .08-.94c0-.12.08-.33.26-.25s.09.25 0 .38-.22.37-.06.59c.16.07.15.17.07.23Z"
-    />
-    <path
-      d="M80.78 216.06v-.23c0-.06.1-.2.11-.15.15.64.64 1.2.35 1.94a8.47 8.47 0 0 1-.46-1.56Z"
-      style={{
-        fill: "#f7e5c0",
-      }}
-    />
-    <path
-      className="eb"
-      d="M81.28 219.18c.28.22.28.58.34.91a1.35 1.35 0 0 1-.34-.91Z"
-    />
-    <path
-      d="m100.16 203.76-.18.47 3.23.07a1.31 1.31 0 0 0 .57-.25c.3.14.53-.16.8-.19s.31.06.26.3a1.37 1.37 0 0 1-1.06.78 27.89 27.89 0 0 1-2.89.08 1.21 1.21 0 0 0-1.28-.05 1 1 0 0 1-1.09-.31c.22.29.5.31.49.63s-.21.14-.34.1a1.89 1.89 0 0 1-1.09-1.53.68.68 0 0 1 1-.08c.67.52.68.52 1.43-.06a.37.37 0 0 1 0-.11Z"
-      style={{
-        fill: "#eecd88",
-      }}
-    />
-    <path
-      d="M98.81 202a47.88 47.88 0 0 1 5.49 0c.11.1.11.22 0 .25a3.25 3.25 0 0 1-.54 0 5.48 5.48 0 0 0-2.28 0c-.78.14-1.57 0-2.35.1-.13.06-.53.11-.32-.35Z"
-      style={{
-        fill: "#a36d22",
-      }}
-    />
-    <path
-      d="m103.48 201.86-2.49-.05a6.78 6.78 0 0 1 2.61-.18Z"
-      style={{
-        fill: "#fef2bd",
-      }}
-    />
-    <path
-      className="bv"
-      d="M103.83 202.92c-.12-.18-.39-.15-.46-.4.64.3 1.14-.28 1.73-.35h.1l.25.15a2 2 0 0 1-1.62.6Z"
-    />
-    <path
-      d="M101.44 202.29v-.16a6 6 0 0 1 2.28.14Z"
-      style={{
-        fill: "#6d3c00",
-      }}
-    />
-    <path className="bv" d="M103.37 202.53H102a1.57 1.57 0 0 1 1.37 0Z" />
-    <path
-      d="M120.27 208.11c.1.08.47 0 .24.28-.7 1-.72 2.26-1.08 3.38-.26.8-.44 1.63-.64 2.45a.19.19 0 0 1-.2.17.89.89 0 0 1-.11-.89 4.11 4.11 0 0 0 .22-.92c.1-.56.34-1.08.43-1.64s.32-1.35.56-2c.11-.32.19-.72.58-.83Z"
-      style={{
-        fill: "#9c610c",
-      }}
-    />
-    <path
-      d="M119.89 217a8.06 8.06 0 0 0-.51 1.67c0 .15 0 .4-.24.38a2.45 2.45 0 0 1 .29-1.51c.14-.72.3-1.44.43-2.17a1.69 1.69 0 0 0-.3-1.31c-.06-.2-.39-.25-.28-.55s.23 0 .34 0c1.03 1.05.55 2.28.27 3.49Z"
-      style={{
-        fill: "#7d450c",
-      }}
-    />
-    <path
-      className="ej"
-      d="M120.27 208.11a5.39 5.39 0 0 0-.81 2.15 2.18 2.18 0 0 1-.24.72.46.46 0 0 1-.2-.33c0-.59.34-1.09.28-1.69a5.73 5.73 0 0 1 .13-2.26c.08-.46-.08-1 .27-1.39a.15.15 0 0 1 .19 0l-.28 2.74c.41 0 .78-.65 1.14 0-.46-.18-.46-.18-.48.06Z"
-    />
-    <path
-      d="M122.08 203.94c.33.75 0 1.53.07 2.29-.11 1-.22 1.9-.39 2.84a.39.39 0 0 1-.06.11c-.16.21-.3.66-.54.46s.21-.33.28-.54 0-.2.07-.29a4.5 4.5 0 0 1 .17-1.93c.06-.61.13-1.23.18-1.84a1.67 1.67 0 0 0-.35-1.25c-.07-.25-.63-.18-.34-.66 0-.06.1-.06.16 0a3.63 3.63 0 0 1 .75.81Z"
-      style={{
-        fill: "#8f5613",
-      }}
-    />
-    <path
-      d="M121.55 197.88c.11 0 .14-.31.31-.15a.17.17 0 0 1 0 .27 5.56 5.56 0 0 0-.63 2.48c-.24 1.26-.34 2.56-.63 3.82-.34-.09-.37-.37-.34-.69 0-.5.17-1 .21-1.49.07-.35.11-.7.17-1.05.14-.73.21-1.48.34-2.21.11-.39.17-.8.57-.98Z"
-      style={{
-        fill: "#9b620e",
-      }}
-    />
-    <path
-      className="ej"
-      d="M121.55 197.88a7.36 7.36 0 0 0-.64 2.67 5.35 5.35 0 0 0-.11.55c-.23-.22-.14-.55-.22-.82a11.11 11.11 0 0 0 .08-1.41 9.72 9.72 0 0 1 0-2.41.79.79 0 0 1 .17-.65h.11v2c.33.1.42-.36.69-.27Z"
-    />
-    <path
-      d="M121.33 203.13h-.16a15.81 15.81 0 0 1 .53-3.35 11.12 11.12 0 0 1-.37 3.35Z"
-      style={{
-        fill: "#b26707",
-      }}
-    />
-    <path
-      className="ej"
-      d="M120.64 202.1c-.06.49-.1 1-.19 1.45a.75.75 0 0 0 .18.75.58.58 0 0 1 0 .48c-.21.06-.31-.17-.47-.25a2.87 2.87 0 0 1 .11-2c.25.04.06-.48.37-.43Z"
-    />
-    <path
-      d="M122.21 198.15a.7.7 0 0 1-.38.61c-.08-.39.15-.5.38-.61Z"
-      style={{
-        fill: "#a55e08",
-      }}
-    />
-    <path
-      d="M79.06 200.18a2.2 2.2 0 0 1-.09-1 .39.39 0 0 1 .5-.36c.24 0 0 .26.06.4a3 3 0 0 1 .23 1.17c0 1.39.22 2.77.19 4.16a2.64 2.64 0 0 1 0 .63c-.05.12-.14.31-.22.21s-.32 0-.48 0a.75.75 0 0 1 .1-.53.65.65 0 0 0-.07-1c.2 0 .37 0 .33-.3-.14-1 0-2-.27-3-.04-.27-.14-.34-.28-.38Z"
-      style={{
-        fill: "#e4a62e",
-      }}
-    />
-    <path
-      d="M79.75 210.66a1.38 1.38 0 0 1-.76-.93c-.25-1.3-.27-2.63-.46-3.94 0-.17-.06-.39.19-.43l.1.66a3 3 0 0 1 .23 1.39c.06.55.12 1.1.2 1.64a1.74 1.74 0 0 0 .63 1.23c.13.22.42.15.55.41-.25.24-.43-.16-.68-.03Z"
-      style={{
-        fill: "#764008",
-      }}
-    />
-    <path
-      className="bc"
-      d="M78.72 205.36a4.16 4.16 0 0 0 0 1.79c0 .86.15 1.72.23 2.58a17.62 17.62 0 0 1-.45-3.07 3.26 3.26 0 0 1 .12-2.07c.11 0 .16-.29.31-.14a.3.3 0 0 1 0 .35c-.04.2-.12.37-.21.56Z"
-    />
-    <path
-      d="m79.2 205.36.32-.49c.18.14 0 .45.24.5.09.11.12.28.25.34s.14.17.08.3a.55.55 0 0 1-.44.11c-.17-.23-.46-.38-.45-.76Z"
-      style={{
-        fill: "#c58a1c",
-      }}
-    />
-    <path
-      d="M79.05 207.41a4.65 4.65 0 0 1-.23-1.39l.06-.55a3.48 3.48 0 0 1 .17 1.94Z"
-      style={{
-        fill: "#f9e4c0",
-      }}
-    />
-    <path
-      d="m94.23 235.75-.46-.24a.15.15 0 0 1 .12-.15c.46-.1.82.25 1.23.31a3.06 3.06 0 0 1 1.81.67 11.48 11.48 0 0 0 1.16.49 2.52 2.52 0 0 1 1.35 1c-.2.13-.37.34-.54-.1-.11-.27-.47-.33-.73-.43Z"
-      style={{
-        fill: "#9a620e",
-      }}
-    />
-    <path
-      d="M97.16 237.49a12 12 0 0 0-1.8-.79c-.12-.06-.27-.08-.35-.18s.06-.09.12-.11l2.12.79c.09 0 .27 0 .17.22s-.17.05-.26.07Z"
-      style={{
-        fill: "#ad6003",
-      }}
-    />
-    <path
-      d="M121.68 206.88a7.8 7.8 0 0 1-.17 1.93h-.16a2.61 2.61 0 0 1 .33-1.93Z"
-      style={{
-        fill: "#f3dcb5",
-      }}
-    />
-    <path
-      className="ap"
-      d="M121.76 209.07a17 17 0 0 1 .39-2.84 14.75 14.75 0 0 1-.3 2.82.15.15 0 0 1-.09.02Z"
-    />
-    <path
-      d="M115.65 224.23c-.23-.37 0-.68.1-1s.15-.43.23-.64a3.86 3.86 0 0 1 .73-1.81 4.64 4.64 0 0 0 .62-1.4 5 5 0 0 1 .8-1.22c.11-.14.16 0 .23 0 0 .16.18.28-.12.46-.57.36-.66 1.15-.94 1.77-.54 1.19-1 2.43-1.5 3.65a.84.84 0 0 1-.15.19Z"
-      style={{
-        fill: "#9b610c",
-      }}
-    />
-    <path
-      d="M118 219.28c.17-.25.19-.5.43-.49 0 0 .1.09.09.11-.04.24-.23.28-.52.38Z"
-      style={{
-        fill: "#a46407",
-      }}
-    />
-    <path
-      d="M108.3 235.25c.16.15.1.35.09.54 0 0-.09 0-.09.07-.11.8-.54.77-1.06.54l.41-.32c-.46-.22-.83 0-1.21.2a23.54 23.54 0 0 0-3.58 1.45c-.28 0-.18-.2-.09-.36s.26-.28.44-.35c1.28-.45 2.54-1 3.82-1.44a1.66 1.66 0 0 1 .73-.08.52.52 0 0 0 .54-.25Z"
-      style={{
-        fill: "#ebad36",
-      }}
-    />
-    <path
-      d="M111.54 233.55a.85.85 0 0 0 .08-.71c-.06-.26 0-.41.25-.41a2.46 2.46 0 0 1-1.12 2.62c-.62.44-1.2 1-1.79 1.46-.19.16-.36.33-.61.29.31-.67 1-.91 1.44-1.35.6-.61 1.39-1.01 1.75-1.9Z"
-      style={{
-        fill: "#804b0f",
-      }}
-    />
-    <path
-      d="M109.79 235.45a9.16 9.16 0 0 1-1.44 1.35h-.29a3.27 3.27 0 0 1 1.73-1.35Z"
-      style={{
-        fill: "#f7e1bb",
-      }}
-    />
-    <path
-      d="M119.43 217.53a8.57 8.57 0 0 1-.29 1.51l-.1.14a2.39 2.39 0 0 1 .32-1.81Z"
-      style={{
-        fill: "#f8e4c1",
-      }}
-    />
-    <path
-      d="M112.64 229.58c.3-.12.46-.43.65-.68a3.18 3.18 0 0 1 1.22-1.09c.19.19.53.35-.07.46s-.71.67-1 1c-.9 1.09-1.77 2.2-2.66 3.31h-.19c-.13-.58.51-.68.56-1.16.3-.66.89-1.05 1.21-1.69.03-.07.11-.22.28-.15Z"
-      style={{
-        fill: "#98600e",
-      }}
-    />
-    <path
-      d="M113.65 229.74a8.57 8.57 0 0 1-1.16 1.38c0 .06-.09 0-.13 0a9 9 0 0 1 1-1.36c.1-.14.19-.1.29-.02Z"
-      style={{
-        fill: "#a75c02",
-      }}
-    />
-    <path
-      d="M112.36 231.11h.13l-1.27 1.67h-.14a6 6 0 0 1 1.28-1.67Z"
-      style={{
-        fill: "#aa6001",
-      }}
-    />
-    <path
-      d="M114.83 228.58c-.2.2-.37.31-.64.21.16-.28.32-.41.64-.21Z"
-      style={{
-        fill: "#ac6508",
-      }}
-    />
-    <path
-      d="M84.11 225.94s.12-.08.13-.07c.33.42.78.78.68 1.46a5.63 5.63 0 0 1-.81-1.39Z"
-      style={{
-        fill: "#f6e4be",
-      }}
-    />
-    <path
-      className="fg"
-      d="M83.58 218.57c-.37-1-.47-2.15-.94-3.21-.23.34-.26.83-.7.85v-.11c0-1-.24-2-.28-3a2 2 0 0 0 0-1.11c-.06-.26-.16-.54.14-.72.22.08.25.3.29.51.07.37.18.73.23 1.11a1.25 1.25 0 0 0 .61 1 2.16 2.16 0 0 1 1 1.4c.25 1 .5 1.92.76 2.88a4.36 4.36 0 0 1-.29 2.53l-.13.16c-.12 0-.18-.14-.23-.25-.07-.27-.24-.51-.14-.81.25-.37.06-.8.12-1.2a.58.58 0 0 0 .08-.63c-.1-.38-.21-.76-.31-1.13 0-.15-.09-.17-.19-.08-.26-.18-.24-.51-.32-.78a2.36 2.36 0 0 0-.59-1.32c.68 1.05.73 2.36 1.11 3.53a.41.41 0 0 1 0 .37.11.11 0 0 1-.22.01Z"
-    />
-    <path
-      d="M81.75 213.15c.09 1 .54 1.93.2 2.95-.23-.1-.51-.13-.25-.55a.88.88 0 0 0 0-.56c-.37-1.83-.67-3.68-1.13-5.48.37-.23.5.07.61.37a11.4 11.4 0 0 1 .25 1.16 9.33 9.33 0 0 1 .32 2.11Z"
-      style={{
-        fill: "#9c620d",
-      }}
-    />
-    <path
-      className="fg"
-      d="M81.35 211.08c-.1-.38-.2-.77-.28-1.16s-.2-.44-.47-.41V209c.36.05.52-.11.49-.53a18.19 18.19 0 0 0-.44-3.3c.21-.15.24.08.33.19a28.51 28.51 0 0 1 .55 4v.05c0 .05.08 0 .1 0a1.49 1.49 0 0 0 .2-.85c.05-.15.06-.36.28-.29a2.43 2.43 0 0 1 0 1.66 5.53 5.53 0 0 0-.24.91c-.03.44-.3.29-.52.24Z"
-    />
-    <path
-      d="M78.58 189.55a1.79 1.79 0 0 1-.24-1.32c.09-.61.21-1.2.31-1.8a5.11 5.11 0 0 0 .37-2.05c.06-.19.21-.27.35-.37.35-.49.35-.49.79-.43 0 .31-.33.24-.41.45s-.66.4-.13.78a.28.28 0 0 1 0 .29 6.21 6.21 0 0 0-.48 1.87 6.58 6.58 0 0 1-.39 1.51c-.34.29.06.75-.17 1.07Z"
-      style={{
-        fill: "#eccc8e",
-      }}
-    />
-    <path
-      d="m122.81 196.25.09.23v1.56l-.11.37a4.27 4.27 0 0 1 .02-2.16Z"
-      style={{
-        fill: "#f3dcb9",
-      }}
-    />
-    <path
-      d="M120.88 184.11c.14-.3.29-.59.61-.68a.23.23 0 0 1 .19.29c-.4.22-.42.79-.74 1.11-.13-.26-.03-.49-.06-.72Z"
-      style={{
-        fill: "#9f671a",
-      }}
-    />
-    <path
-      d="M110.29 233.13c-.25 0-.18-.39-.34-.52.19-.15.37-.49.55 0 .09.25-.07.39-.21.52Z"
-      style={{
-        fill: "#f4e6bd",
-      }}
-    />
-    <path
-      d="M84.57 194.15c0 .34 0 .69.37.86l-.4.07c-.47.12-.75-.26-1-.56a10.1 10.1 0 0 1-1.07-1.62 3.22 3.22 0 0 0-1.2-1.84c-.28-.32 0-.85-.31-1.14a.48.48 0 0 1 0-.26.44.44 0 0 1 .07-.17c.1-.11.19-.23.35-.16.36 0 .45.19.39.55a3.55 3.55 0 0 0-.06.56c0 .13 0 .32.14.33s.13-.17.19-.26a.18.18 0 0 1 .13.24l.1.26c-.23.24 0 .39.11.53.48.77 1 1.55 1.49 2.29.13.24.33.59.7.32Z"
-      style={{
-        fill: "#ebcb88",
-      }}
-    />
-    <path
-      d="M100 203.72c-.52.76-.85.84-1.43.24-.31-.32-.62-.21-.95-.16a.24.24 0 0 1-.09-.13c0-.15.12-.34-.08-.45-.86-.34-1.76-.57-2.63-.88-.36-.15-.74-.21-1.09-.35a.56.56 0 0 0-.85.39.42.42 0 0 1-.16 0 .33.33 0 0 1-.34-.25c0-.32.28-.1.33-.28a.61.61 0 0 1-.08-.87.59.59 0 0 0 .44.26l5.59 1.88a1.26 1.26 0 0 1 .74.44c.22.3.41.25.6.16Z"
-      style={{
-        fill: "#e6aa2e",
-      }}
-    />
-    <path
-      d="M92.74 201.89c-.08.09-.25.24-.23.26.12.21.3.14.48.09a.23.23 0 0 1 .41 0c.11.24-.11.3-.25.38a1.53 1.53 0 0 1-1.11-.12 11 11 0 0 1-1.94-1.18 2.64 2.64 0 0 0-1.69-1.13.26.26 0 0 1-.25-.18c0-.39.07-.57.44-.43a.4.4 0 0 0 .32 0l.8.06-.49.37c1 .61 1.87 1.19 2.79 1.76a1.19 1.19 0 0 0 .72.12Z"
-      style={{
-        fill: "#ebcd8c",
-      }}
-    />
-    <path
-      d="M80.11 198.91a20 20 0 0 0-.19-3.78c-.24.29-.36.68-.77.57a.41.41 0 0 1-.05-.18c.52-2 .23-4 .37-6V188.95c.75-.45.81-1.24.82-2.06a.45.45 0 0 1 .31-.3c.11.05.09.17.11.27.21.4 0 .78 0 1.16-.05.53-.24 1-.28 1.56.15-.06.23-.22.34-.32s.18-.18.29-.06a.27.27 0 0 1 0 .34c-.13.22-.18.5-.39.66a1.27 1.27 0 0 0-.4 1v1a2.18 2.18 0 0 0 .52 1.8 2.84 2.84 0 0 1 .54 1.91c0 .92-.05 1.83 0 2.76a2.45 2.45 0 0 1-.69 1.77c-.15.15-.11.55-.35.5s-.21-.41-.27-.64a5.3 5.3 0 0 1 .09-1.39Z"
-      style={{
-        fill: "#eecd83",
-      }}
-    />
-    <path
-      className="fq"
-      d="M81 189.58c-.05-.08.11-.47-.18-.19s-.25.42-.51.43a10 10 0 0 1 .64-2.41c.49 0 .25.4.31.64a1.92 1.92 0 0 1 .1.89v.39Z"
-    />
-    <path
-      d="M80.62 186.63c0 .15-.08.24-.23.23a2.73 2.73 0 0 1-.78-1.8v-.25a2 2 0 0 1 1.01 1.82Z"
-      style={{
-        fill: "#be7915",
-      }}
-    />
-    <path
-      d="M81 187.41a10.32 10.32 0 0 1-.64 2.41c-.09 0-.16.17-.25.08s0-.14 0-.22a10.16 10.16 0 0 0 .66-2.78c.29.03.12.33.23.51Z"
-      style={{
-        fill: "#855211",
-      }}
-    />
-    <path
-      d="M81.37 188.94c-.29-.26-.14-.58-.1-.89.36.1.34.5.48.78a.25.25 0 0 1-.13.31c-.13.06-.15-.14-.25-.2Z"
-      style={{
-        fill: "#be7a14",
-      }}
-    />
-    <path
-      d="m82.14 190.75-.13-.24c0-.16 0-.35.16-.44s.11 0 .13.08a.48.48 0 0 1-.16.6Z"
-      style={{
-        fill: "#9c5b05",
-      }}
-    />
-    <path
-      d="M85.7 195.07a1.8 1.8 0 0 0-.44-.47c-.1-.11-.35-.29 0-.44a59.65 59.65 0 0 1 3.92 4.71c-.36.12-.45-.3-.67-.45a2.84 2.84 0 0 1-.47-.59 11.52 11.52 0 0 1-2-2.37Z"
-      style={{
-        fill: "#eeaf33",
-      }}
-    />
-    <path
-      d="M82.54 192.83c.4.56.79 1.13 1.2 1.68a1.58 1.58 0 0 0 .8.57c.37.24.75 0 1.13 0 0 .33-.19.36-.42.35h-1c-.46.13-.74-.22-1-.53a7 7 0 0 1-.85-1.27c-.3-.57-.63-1.11-1-1.66-.13-.22-.41-.45-.09-.75a4.24 4.24 0 0 1 1.23 1.61Z"
-      style={{
-        fill: "#723f09",
-      }}
-    />
-    <path
-      className="fq"
-      d="M82.54 193.61c.52.69.87 1.59 1.76 1.83.24.24.54.09.8.16s.42-.11.57.14h-1.25a1.64 1.64 0 0 1-1.28-.71c-.74-1-1.47-2.1-2.16-3.19a1.59 1.59 0 0 1-.27-1.37.31.31 0 0 0-.06-.35c.1-.19.15-.41.36-.5v.26c-.31 1.65.89 2.54 1.53 3.73Z"
-    />
-    <path
-      className="ad"
-      d="M87.88 200.28a1.54 1.54 0 0 1-.2-1.82c.13.11.39-.27.45.15-.07.08-.18.14-.17.29.07-.06.11-.16.2-.19.48.06.5.61.76.9 0 .32-.19.35-.39.17s-.19.17-.31.23-.19-.21-.32-.27c-.01.18.1.37-.02.54Z"
-    />
-    <path
-      d="m86 195.46 2 2.37c-.13.18-.22.06-.32 0l-1.67-2c-.01-.12-.16-.24-.01-.37Z"
-      style={{
-        fill: "#b26b05",
-      }}
-    />
-    <path
-      d="M88.16 198.71c-.07.17.07.45-.21.51-.12 0-.22-.06-.21-.22s.1-.45.39-.39Z"
-      style={{
-        fill: "#f1dda1",
-      }}
-    />
-    <path
-      className="ag"
-      d="M115.25 198.12a2.07 2.07 0 0 1-.65 2.51c-.93.73-1.88 1.44-2.83 2.14a2.18 2.18 0 0 1-2.8-.14.42.42 0 0 1 .41-.52 2.06 2.06 0 0 1 .6.08c.48.08.62-.05.67-.61.07-.12.16-.16.28-.07s.08.29.26.16c.81-.62 1.64-1.21 2.43-1.86.47-.39.43-.55-.12-.8V199a1.45 1.45 0 0 0 .54-.11c.14.1.22 0 .33-.1.35-.33.48-.14.5.28a2 2 0 0 1-.8 1.26 13 13 0 0 1-1.59 1.2 2.14 2.14 0 0 0-1.54 1.14c-.11.23-.3.23-.46.3a32.65 32.65 0 0 0 3.9-2.65 1.45 1.45 0 0 0 .68-1.31c.03-.34-.06-.66.19-.89Z"
-    />
-    <path
-      d="M109.43 202.24c-.28-.06-.32.24-.46.39a15.68 15.68 0 0 0-3.45 1c-.16-.34 0-.46.27-.54a13.08 13.08 0 0 1 2-.62 1.51 1.51 0 0 0 .72-.37 1.36 1.36 0 0 1 .77-.35s-.09 0-.1-.11c.2-.49.6-.31.9-.36-.19-.06-.45.1-.64-.15.4-.27.88.09 1.27-.24.18.18.06.47.2.66-.09 0-.18 0-.23.12a.58.58 0 0 1-.71.54 1.08 1.08 0 0 0-.54.03Z"
-      style={{
-        fill: "#cf9224",
-      }}
-    />
-    <path
-      className="bv"
-      d="m112.36 198.31-.11.27-2 1.55a.5.5 0 0 0-.26.46c-.12.09-.25.11-.34 0s0-.18.1-.25l1.31-1.1Z"
-    />
-    <path
-      className="ag"
-      d="M109.7 200.32c-.11.31.07.28.24.27h.44a5.57 5.57 0 0 1-1.4.2c.28-.07.42-.41.72-.47Z"
-    />
-    <path
-      className="ga"
-      d="M114.08 198.89a.71.71 0 0 1-.57.06c.36-.07.29-.33.23-.6l4-4.35a.78.78 0 0 1 .82-.24c0-.3-.14-.36-.26-.47a.23.23 0 0 1 .14-.25c.53-.08.53.4.63.77.07.81-.54.84-1 1.1a1.1 1.1 0 0 0-.2.35.3.3 0 0 1-.09.13c-.12.06-.25 0-.36.08a.18.18 0 0 1-.1 0 .6.6 0 0 0-.25-.1c-.3.07-.46.36-.67.57l-1.85 2a1.13 1.13 0 0 0-.38.69.48.48 0 0 1-.09.26Z"
-    />
-    <path
-      d="M119.63 189.52h.13c.22-.08.25-.34.36-.52s.1-.42.36-.39.35.28.34.56c-.19.4-.4.78-.85.83s-.39-.18-.34-.48Z"
-      style={{
-        fill: "#cf9120",
-      }}
-    />
-    <path
-      d="M120 190c.41-.11.46-.66.81-.84.28.08.36.29.29.6s-.15.66-.45.82c-.26-.09 0-.41-.21-.52a.23.23 0 0 0-.32 0Z"
-      style={{
-        fill: "#edbd54",
-      }}
-    />
-    <path
-      d="M109.46 201.09c.39.09.83-.12 1.31.29-.6.06-1.13-.15-1.56.22a1.06 1.06 0 0 1-.65.37c-1.41.41-2.8.91-4.21 1.34-.18.05-.44.24-.55-.14v-.25l1.62-.6a13.26 13.26 0 0 0 2.28-.76 10.28 10.28 0 0 0 1.76-.47Z"
-      style={{
-        fill: "#9a5f0b",
-      }}
-    />
-    <path
-      d="M121 189.77a.91.91 0 0 0-.24-.64c-.08-.2-.34-.26-.32-.53s.32-.2.49 0 .19.4.3.58a2.09 2.09 0 0 1 0 2c-.1.2-.17.42-.25.63-.34.85-.77 1.65-1.13 2.49-.09.21-.25.62-.5.09a12.88 12.88 0 0 1 .6-1.49c.28-1.08.84-2.03 1.05-3.13Z"
-      style={{
-        fill: "#77400a",
-      }}
-    />
-    <path
-      d="M121.25 189.13c-.31-.09-.28-.81-.79-.53-.2.17-.2.45-.31.66s-.11.48-.39.23c.57-1.22 1.13-2.44 1.7-3.65a3.81 3.81 0 0 1 1-1.7l-.61 1.4a3.41 3.41 0 0 1-.61 1.42c-.23.25 0 .63-.19.89-.28.09-.15.31-.1.48s.25.25.26.46Z"
-      style={{
-        fill: "#e9ad34",
-      }}
-    />
-    <path
-      d="M121.28 187a3.92 3.92 0 0 1 .61-1.42c.29.11.26-.21.34-.34.17-.31.3-.65.64-.78l-.15.51-1.19 2.41c-.22.1-.22.1-.25-.38Z"
-      style={{
-        fill: "#c27d1a",
-      }}
-    />
-    <path
-      d="M120.78 195.81a12 12 0 0 0 0 3.05c-.31.19-.37-.08-.45-.31v-3.51a.93.93 0 0 1 .16-.54v-.24c.1 0 0-.33.23-.25Z"
-      style={{
-        fill: "#895107",
-      }}
-    />
-    <path
-      className="ga"
-      d="M119.42 194.36c.09.39.19.11.25 0 .38-.76.75-1.52 1.13-2.29a1.84 1.84 0 0 1 .17-.25 3.86 3.86 0 0 1-.48 1.29c-.09.26-.28.47-.25.78v.11c-.12.59-.58.82-.94 1.17-.24.05-.53.39-.71.05s.2-.52.4-.73a.43.43 0 0 1 .43-.13Z"
-    />
-    <path
-      d="M120.24 193.87c.07 0 .2-.33.22 0v.62a11.64 11.64 0 0 1 0 1.67.53.53 0 0 1 0 .12c-.08.25.11.74-.22.72s-.14-.51-.12-.79a7.85 7.85 0 0 1 .15-2.37Z"
-      style={{
-        fill: "#fdf0cb",
-      }}
-    />
-    <path
-      className="gj"
-      d="M120.4 196.19c0-.56 0-1.11.08-1.66h.08c0 1.38 0 2.76-.08 4.14-.45-.27-.28-.76-.24-1.13s.1-.91.16-1.35Z"
-    />
-    <path
-      d="M121.25 188.79c-.51-.6-.53-.71-.26-1.27.12.07.07.22.1.33-.12.37.09.55.29.78s.07.25-.13.16Z"
-      style={{
-        fill: "#976523",
-      }}
-    />
-    <path
-      className="gj"
-      d="m120.81 194-.23.25c0-.66 0-.67.39-1.11.19.37-.08.6-.16.86Z"
-    />
-    <path
-      d="M123.28 184.31v1.61a2.85 2.85 0 0 1 0-1.61Z"
-      style={{
-        fill: "#fef3d1",
-      }}
-    />
-    <path
-      d="M117.05 214.6a1.66 1.66 0 0 1 .83-1.3.79.79 0 0 0 .46-.53 15.33 15.33 0 0 0 .65-3.22c-.07-.18.08-.25.15-.35s.17-.21.31-.21a8.86 8.86 0 0 1-.36 1.7 6.13 6.13 0 0 1-.34 1.65c0 .1-.12.23.07.27 0 .61-.46 1.14-.26 1.79-.1 0-.18-.22-.33.08s-.26 1-.61 1.34c-.25-.44.09-.83.1-1.25 0-.23.14-.43.15-.65-.23 0-.25.26-.36.4s-.14.45-.34.57c-.1-.07-.1-.18-.12-.29Z"
-      style={{
-        fill: "#edd088",
-      }}
-    />
-    <path
-      d="M108.3 235.25c0 .36-.09.58-.53.42s-1 .19-1.41.35c-1.2.42-2.39.9-3.59 1.35-.11-.06-.12-.18-.13-.29.39-.33.9-.31 1.31-.59a4 4 0 0 1 1.93-.76 12.53 12.53 0 0 1 2.42-.57Z"
-      style={{
-        fill: "#9c600d",
-      }}
-    />
-    <path
-      d="M114.36 224.49c.24-.45.37-1 1-1.25a16.55 16.55 0 0 0-1.23 3.15c-.42.58-.45 1.37-.89 2.12-.19-1 .4-1.53.44-2.25a3.58 3.58 0 0 1 .68-1.77Z"
-      style={{
-        fill: "#9c661d",
-      }}
-    />
-    <path
-      d="M98.43 236.43a21.76 21.76 0 0 1 4.49-.19v.28Z"
-      style={{
-        fill: "#6d3d01",
-      }}
-    />
-    <path
-      d="M117.48 216.56v.26a6.35 6.35 0 0 0-.47 2.62 1.07 1.07 0 0 1-.31-1.08l.33-1.67a.52.52 0 0 1 .19-.67c.23.09.1.41.26.54Z"
-      style={{
-        fill: "#a76e20",
-      }}
-    />
-    <path
-      d="M108.55 233.48a11.55 11.55 0 0 1-1.87 1.56c.22-.84 1-1 1.42-1.7 0-.26.16-.22.31-.19a.31.31 0 0 1 .14.33Z"
-      style={{
-        fill: "#ab6f20",
-      }}
-    />
-    <path
-      d="m92.55 233.56-1.89-1.33.09-.17c.7 0 1.11.71 1.71.95a.36.36 0 0 1 .36.39.33.33 0 0 1-.27.16Z"
-      style={{
-        fill: "#ab7020",
-      }}
-    />
-    <path
-      d="M108.55 233.48c-.1-.07-.1-.19-.12-.3a3 3 0 0 1 1.3-1.22c.24 0 .38-.26.69-.18a14.81 14.81 0 0 1-1.87 1.7Z"
-      style={{
-        fill: "#875516",
-      }}
-    />
-    <path
-      d="M92.79 233.41c0-.21-.25-.23-.33-.4.1-.15.23-.13.35 0 .46.33.91.67 1.36 1 .09.07.22.2.07.36a.45.45 0 0 1-.39-.07c-.37-.3-.8-.45-1.06-.89Z"
-      style={{
-        fill: "#f9eabb",
-      }}
-    />
-    <path
-      d="m89.83 231.57-1.42-2c0-.2.09-.21.21-.08l.8.88c0 .34.31.52.43.81s.15.28-.02.39Z"
-      style={{
-        fill: "#f9eab4",
-      }}
-    />
-    <path
-      d="m117.25 216-.22.66c-.12.08-.29.27-.22-.1.1-.57.24-1.13.36-1.7a1.78 1.78 0 0 1 .25-.64c.19 0 .18.15.19.29-.11.49.05 1.11-.36 1.49Z"
-      style={{
-        fill: "#fae9be",
-      }}
-    />
-    <path
-      d="M109.73 232c-.43.41-.86.82-1.3 1.22l-.33.16c-.09 0-.22-.05-.1-.18a7.68 7.68 0 0 1 1.69-1.39c.16-.13.03.08.04.19Z"
-      style={{
-        fill: "#fcebbe",
-      }}
-    />
-    <path
-      d="M112.64 229.58a9.07 9.07 0 0 1-1.52 1.89.14.14 0 0 1-.06-.21 14.21 14.21 0 0 1 1.35-1.72c.08-.08.17-.07.23.04Z"
-      style={{
-        fill: "#f7eab7",
-      }}
-    />
-    <path
-      d="M105.88 235.73a11.38 11.38 0 0 1-1.93.76l-.32-.07a6.22 6.22 0 0 1 2.13-.85c.09-.03.11.07.12.16Z"
-      style={{
-        fill: "#faedbb",
-      }}
-    />
-    <path
-      d="m96.93 236.34-1.81-.67c.4-.37.73 0 1.09.11a7.52 7.52 0 0 1 .78.32c.24.13.08.2-.06.24Z"
-      style={{
-        fill: "#fbeebd",
-      }}
-    />
-    <path
-      d="M116.71 220.77a10.54 10.54 0 0 1-.73 1.81c-.11 0-.27.07-.19-.13a9.47 9.47 0 0 1 .77-1.82s.06-.19.14 0a.59.59 0 0 1 .01.14Z"
-      style={{
-        fill: "#f9eaba",
-      }}
-    />
-    <path
-      d="m92.79 233.41 1.13.79a1.21 1.21 0 0 1 .33.38c0 .11 0 .21-.12.15l-1.58-1.17a.21.21 0 0 1 .24-.15Z"
-      style={{
-        fill: "#855417",
-      }}
-    />
-    <path
-      d="M114.36 224.49a12 12 0 0 1-.73 1.77c-.2-.55.16-1 .29-1.45a3.12 3.12 0 0 1 .23-.49c0-.08.09-.18.15-.12s.04.19.06.29Z"
-      style={{
-        fill: "#faedc3",
-      }}
-    />
-    <path
-      d="m93.89 235.36-.12.15-.44-.13c0-.39.3-.34.44-.52.23.14.4.23.12.5Z"
-      style={{
-        fill: "#f5e3bd",
-      }}
-    />
-    <path
-      d="M102.19 236h-1.9v-.07h1.91Z"
-      style={{
-        fill: "#fef9cd",
-      }}
-    />
-    <path
-      d="M94.13 234.73c.15-.29-.22-.31-.21-.53l.32.15c.12.24.37.35.45.68a.72.72 0 0 1-.56-.3Z"
-      style={{
-        fill: "#ad7a38",
-      }}
-    />
-    <path
-      d="M85.41 225.54c-.25.11-.46.29-.36-.28.13-.76-.36-1.36-.57-2-.42-1.28-.9-2.53-1.36-3.79.33-.11.49.14.61.41a8.49 8.49 0 0 1 .36 1 6.16 6.16 0 0 1 .59 1.85c.15.74.57 1.38.67 2.15.03.2.02.43.06.66Z"
-      style={{
-        fill: "#99600f",
-      }}
-    />
-    <path
-      d="M86.88 226.21a23 23 0 0 0-1.49-2.21c.5.09.5.09 1.5 1.62a1 1 0 0 1 .49.87c-.27.15-.31-.26-.5-.28Z"
-      style={{
-        fill: "#ab7023",
-      }}
-    />
-    <path
-      d="M86.88 226.21c.24-.11.27.25.45.27a1.09 1.09 0 0 1 .62 1.07 1.86 1.86 0 0 1 .3 1.16Z"
-      style={{
-        fill: "#814e12",
-      }}
-    />
-    <path
-      d="m88 227.55-.62-1.07c0-.37-.4-.54-.44-.91 0-.07.14-.13.19-.06a7.49 7.49 0 0 1 1.1 1.82c.08.29-.13.24-.23.22Z"
-      style={{
-        fill: "#f7e5b7",
-      }}
-    />
-    <path
-      d="m84.68 222.7-.68-1.81c-.06-.14-.09-.28.13-.26l.11.26c.21.5.42 1 .6 1.5.05.09.22.43-.16.31Z"
-      style={{
-        fill: "#fbecbd",
-      }}
-    />
-    <path
-      className="dg"
-      d="M102.54 237.09c.07.36-.22.45-.38.64-.35-.57.34-.36.38-.64Z"
-    />
-    <path
-      className="aq"
-      d="M80.44 200.37c0 1 0 2 .08 2.94 0 .56.55.59.75 1 .11.22 0 .26-.16.23s-.21-.18-.35-.09 0 .3.11.44.21.35 0 .54c-.1 0-.11-.23-.25-.23-.21.27-.13.78-.56.85l-.13-.26c-.06-1 0-1.91-.07-2.87s-.07-1.79-.15-2.68a.9.9 0 0 1 .14-.72c.16-.15 0-.48.23-.57.15.62-.15 1.27.1 1.88l.05.1c.16-.19.02-.44.21-.56Z"
-    />
-    <path
-      d="M79.52 188.93a1.15 1.15 0 0 1 .11.73c0 1.52-.09 3-.1 4.54a8 8 0 0 1-.38 1.5h-.09a.63.63 0 0 1 0-.65 38.2 38.2 0 0 0 .23-4.41 5 5 0 0 1 .14-1.76s.07 0 .09.05Z"
-      style={{
-        fill: "#a06311",
-      }}
-    />
-    <path
-      d="M79.88 199.48c0 1.38.13 2.75.21 4.13a4.55 4.55 0 0 1-.13 2.14c-.37.07-.19-.22-.2-.38a53.26 53.26 0 0 0-.26-5.93 1 1 0 0 1 0-.24Z"
-      style={{
-        fill: "#9a5f0c",
-      }}
-    />
-    <path
-      d="M78.84 190.3h.09a10.54 10.54 0 0 1 0 3 12.79 12.79 0 0 1-.09-3Z"
-      style={{
-        fill: "#b36a00",
-      }}
-    />
-    <path
-      d="M79 184.38a2.22 2.22 0 0 1-.37 2.05 7.42 7.42 0 0 1 .37-2.05Z"
-      style={{
-        fill: "#f2ddb8",
-      }}
-    />
-    <path
-      className="dh"
-      d="M84.14 220.63c-.1.05-.11.16-.13.26-.14-.34-.29-.68-.42-1a.68.68 0 0 0-.47-.42s0-.07 0-.11c.48 0 .41-.42.45-.76h.11a1.06 1.06 0 0 1 .25 1.26c-.03.27.21.48.21.77Z"
-    />
-    <path
-      d="M84.75 224.76c.09.22.21.38.07.56a.17.17 0 0 1-.15 0 .52.52 0 0 1 .08-.56Z"
-      style={{
-        fill: "#a25d07",
-      }}
-    />
-    <path
-      d="M103.8 203.17c.43.17.79-.12 1.17-.24 1.42-.42 2.83-.88 4.24-1.33l.38.06-1 .47c-.94.41-1.92.62-2.87 1a1.24 1.24 0 0 0-.37.17c-.31.16-.67.17-.93.44v.26c-.24 0-.41.72-.72.09a1.21 1.21 0 0 1-.5-.66Z"
-      style={{
-        fill: "#e9aa32",
-      }}
-    />
-    <path
-      className="hs"
-      d="M105.29 203.18a18.65 18.65 0 0 1 3.33-1.05.81.81 0 0 1-.66.44c-.75.11-1.43.46-2.16.64-.22 0-.29.15-.28.38a2.09 2.09 0 0 1-1.83 1.73 26.46 26.46 0 0 1-4.71 0c-.12-.24-.38-.26-.54-.42s-.24-.14-.15-.28a.2.2 0 0 1 .34 0 .84.84 0 0 0 .92.31c.27-.05.53-.18.79 0-.16.4-.42.19-.66.15v.09c1-.11 1.93 0 2.89-.12a2.62 2.62 0 0 1 .74 0 1.7 1.7 0 0 0 1.48-.67.65.65 0 0 0 .33-.61.69.69 0 0 1 .17-.59Z"
-    />
-    <path
-      d="M104.92 204.33a2.32 2.32 0 0 1-2.28.76 5.17 5.17 0 0 1-1.84-.08v-.12c.94 0 1.89-.1 2.84-.09a1.54 1.54 0 0 0 1.14-.73.14.14 0 0 1 .17 0 .19.19 0 0 1-.03.26Z"
-      style={{
-        fill: "#5f340c",
-      }}
-    />
-    <path
-      className="hs"
-      d="M104.94 204.08h-.16c-.06-.15-.18-.11-.28-.11s-.45.11-.25-.29c.14 0 .23-.15.36-.18s.34-.22.47 0a.43.43 0 0 1-.14.58Z"
-    />
-    <path
-      className="do"
-      d="m100.34 204.93-.79-.06c.42-.22.84-.44 1.25 0a.37.37 0 0 1-.46.06Z"
-    />
-    <path
-      d="m117.25 216 .27-1.54c.1-.34.14-.72.51-.86.22.49-.28.8-.24 1.26a3.89 3.89 0 0 1-.17.93l-.14.74c-.33-.01-.17-.31-.23-.53Z"
-      style={{
-        fill: "#855212",
-      }}
-    />
-    <path
-      d="m119.24 209.27-.25.28a3.38 3.38 0 0 1-.27-2.81 17.76 17.76 0 0 1 .35-2.52c.11-.31.2-.67.42-.8.67-.4.65-1.15.8-1.8a13.67 13.67 0 0 0 .17-1.45c.21 0 .19.21.24.36a8 8 0 0 1-.31 1.95 6.18 6.18 0 0 0-.26 2.05c-.28.18-.21.5-.24.77h-.12a1.55 1.55 0 0 1 0-1.16c0-.1.18-.22 0-.3a1.93 1.93 0 0 0-.4 1c-.16.49 0 1-.29 1.61.14-.16.09-.44.32-.28.17.31 0 .62 0 .92a9 9 0 0 0-.16 2.18Z"
-      style={{
-        fill: "#f1d190",
-      }}
-    />
-    <path
-      d="M118.82 212.61c-.1.06-.28.32-.21-.09a9.29 9.29 0 0 1 .48-1.83c.11.06.1.18.13.29a11.4 11.4 0 0 1-.4 1.63Z"
-      style={{
-        fill: "#faeec2",
-      }}
-    />
-    <path
-      d="M119.71 203.79c.09-.07.19-.22.26-.2s.09.22 0 .31c-.28.43-.16.94-.24 1.4a16.89 16.89 0 0 0-.32 3.69l-.21.28c-.43-.78-.06-1.59 0-2.38a3.17 3.17 0 0 0 .07-.69 1.88 1.88 0 0 1-.07-1c.2-.44.14-1.02.51-1.41Z"
-      style={{
-        fill: "#9f691f",
-      }}
-    />
-    <path
-      d="M120.39 202.48a7.52 7.52 0 0 1 .29-2s-.05-.06 0-.08c.19.23 0 .53.11.78a3.67 3.67 0 0 1-.16 1c-.18-.04.01.31-.24.3Z"
-      style={{
-        fill: "#f7ebb9",
-      }}
-    />
-    <path
-      className="cy"
-      d="M120.69 200.32v.11l-.22-.26a1 1 0 0 0-.14-1 2.2 2.2 0 0 1-.42-1.54c0-.91 0-1.82.07-2.74-.31-.1-.41.33-.67.28a8.26 8.26 0 0 0 .94-1.26l-.06 3.12c.2-.27 0-.6.22-.8v1.78c0 .23-.06.47.11.67s.12.34.33.22a4.49 4.49 0 0 1-.16 1.42Z"
-    />
-    <path
-      d="M83.94 219.83a3.58 3.58 0 0 0-.25-1.26c-.37-1.28-.4-2.71-1.23-3.8a.29.29 0 0 1 0-.35c.1-.12.21 0 .3.06a1.47 1.47 0 0 1 .46.78 12.61 12.61 0 0 0 .42 1.52 2.78 2.78 0 0 1 .42 1.81 3 3 0 0 1 .12 1c-.03.15-.05.28-.24.24Z"
-      style={{
-        fill: "#99641c",
-      }}
-    />
-    <path
-      d="M81.73 210.93a.49.49 0 0 1 .09.39c-.26.26 0 .51 0 .76s.21.73-.07 1.07c-.13-.69-.27-1.38-.4-2.07l.05-.28c.15-.06.09.44.33.13Z"
-      style={{
-        fill: "#fbedbc",
-      }}
-    />
-    <path
-      d="m84.05 218.59-.42-1.81c0-.07.08-.28.15-.14a5 5 0 0 1 .51 1.84c.02.23-.15.1-.24.11Z"
-      style={{
-        fill: "#f9ecc1",
-      }}
-    />
-    <path
-      d="M80.9 205.39a.67.67 0 0 0-.21-.53c-.15-.17-.4-.44 0-.64.59.22.72.79.78 1.39s.09 1.26.18 1.88.33 1.34.07 2l-.26.55Z"
-      style={{
-        fill: "#a97629",
-      }}
-    />
-    <path
-      d="M81.53 207.48c-.07-.67-.15-1.35-.21-2a1.66 1.66 0 0 0-.61-1.24c.17-.2.25 0 .35.1s.18.18.21 0a2.39 2.39 0 0 1 .58 1.53c0 .82.17 1.63.26 2.45-.14.09-.12.25-.14.39-.4-.29-.19-.78-.3-1.17v-.06c0 .02-.1-.02-.14 0Z"
-      style={{
-        fill: "#f1d390",
-      }}
-    />
-    <path
-      d="M81.53 207.48a2.26 2.26 0 0 1 .17-.83c.16.73.08 1.39.27 2a1.09 1.09 0 0 1-.23.88Z"
-      style={{
-        fill: "#fcecbb",
-      }}
-    />
-    <path
-      className="ab"
-      d="M82.54 193.61a10.09 10.09 0 0 1-1.61-2.35 1.66 1.66 0 0 1 .07-1.41c.54.22.19.77.33 1.14a.2.2 0 0 1 .08.24c-.24.32.07.48.18.68a11.24 11.24 0 0 1 .95 1.7Z"
-    />
-    <path
-      d="m81.41 191.23-.08-.24a3 3 0 0 1 1.17 1.88 8.24 8.24 0 0 1-1.09-1.64Z"
-      style={{
-        fill: "#f8e7c5",
-      }}
-    />
-    <path
-      d="M97.62 203.7a29.85 29.85 0 0 0-3.12-1.13c-.16-.06-.25.05-.36.13a2.21 2.21 0 0 1-2.59.13 22.09 22.09 0 0 1-3.25-2.16c.41-.34.64.12.94.27.76.48 1.51 1 2.26 1.47a2.25 2.25 0 0 0 2.2.17c-.18.06-.35.18-.51 0a1.63 1.63 0 0 0 .15-.2c-.09-.13-.23-.11-.35-.1-.14-.41 0-.59.36-.51a5.77 5.77 0 0 1 1.46.46 13 13 0 0 0 2.24.81 1.42 1.42 0 0 0 .61.16c.42.17-.16.31-.04.5Z"
-      style={{
-        fill: "#c88a23",
-      }}
-    />
-    <path
-      d="M97.66 203.17c-.16.22-.33 0-.47 0-.78-.21-1.57-.45-2.35-.7-.12 0-.39-.11 0-.25a16.23 16.23 0 0 1 2.82.95Z"
-      style={{
-        fill: "#a76202",
-      }}
-    />
-    <path
-      className="bx"
-      d="M99.27 204.12a1 1 0 0 1-.5-.64c.29.14.57.14.5.64Z"
-    />
-    <path
-      d="M93.19 202.55c.23.07.38-.12.57-.22s.25-.21.33 0 0 .28-.19.34a2.33 2.33 0 0 1-2.19 0c-.85-.51-1.66-1.09-2.48-1.63-.14-.13-.21-.25 0-.33a1.37 1.37 0 0 1 .87.52c.5.3 1 .59 1.49.9a2.22 2.22 0 0 0 1.6.42Z"
-      style={{
-        fill: "#7e470f",
-      }}
-    />
-    <path
-      d="M89.23 200.79c-.14.1 0 .19 0 .28-.34 0-.52-.56-.93-.4-.21 0-.3-.23-.42-.39v-.07c0-.27-.4-.62-.1-.78s.23.43.44.58l.2.16c.31.15.69.17.81.62Z"
-      style={{
-        fill: "#b8781b",
-      }}
-    />
-    <path
-      d="m89.23 200.79-.81-.62a2 2 0 0 1 1.71 1.09Z"
-      style={{
-        fill: "#f7e3c2",
-      }}
-    />
-    <path
-      d="M80.61 198.2c-.09 1.35-.09 1.35-.29 1.66 0-1.46-.07-2.94-.11-4.41a1.06 1.06 0 0 0-.23-.83c-.14-.14-.17-.4 0-.5s.29.13.38.29a1.86 1.86 0 0 1 .19.85 16 16 0 0 0 .07 2.3.52.52 0 0 1-.01.64Z"
-      style={{
-        fill: "#965d11",
-      }}
-    />
-    <path
-      d="M80.61 198.2v-.64c0-.36.05-.71.07-1.06h.12l-.06 2.66c-.13-.27.09-.66-.13-.96Z"
-      style={{
-        fill: "#fcedc6",
-      }}
-    />
-    <path
-      className="cx"
-      d="M114.08 198.89v-.28a8.68 8.68 0 0 1 1.34-1.55 15.57 15.57 0 0 1 1.8-1.83c.11.05.12.16.13.28-.63.68-1.26 1.36-1.88 2.05-.13.15-.4.25-.23.56a5.88 5.88 0 0 1-.1 1.39 1.25 1.25 0 0 1-.77.93c-.06-.26.13-.39.24-.55a2.13 2.13 0 0 0 .38-1.28c0 .2 0 .41-.22.43 0-.11 0-.29-.1-.34s-.21.14-.29.25-.28.46-.3-.06Z"
-    />
-    <path
-      className="cy"
-      d="M119.09 194.47a.54.54 0 0 0 0 .11c-.08.23-.5.25-.38.53s.42-.08.64 0a.74.74 0 0 1-.66.26 1.24 1.24 0 0 1-.85 0v-.11c0-.13.5-.08.12-.38a.61.61 0 0 1 .44-.25.7.7 0 0 0 .6-.77c.22 0 .22.2.23.38a.55.55 0 0 1-.14.23Z"
-    />
-    <path
-      className="in"
-      d="M117.79 195.34h.85c-.39.21-.79.05-1.18.08.08-.11.22-.06.33-.08Z"
-    />
-    <path
-      d="M114.78 199c0-.1.07-.19.09-.29s-.23-.54.12-.55.1.34.12.53a2.24 2.24 0 0 1-.73 1.71c-1 .76-2.09 1.54-3.14 2.29a1.15 1.15 0 0 1-1.29.13c.34-.32.77.11 1-.31a3.58 3.58 0 0 1 1.5-1.13c.53-.42 1-.82 1.57-1.25a2 2 0 0 0 .76-1.13Z"
-      style={{
-        fill: "#753e0b",
-      }}
-    />
-    <path
-      d="M112.43 201.42a7.76 7.76 0 0 1-1.5 1.13c-.07 0-.25-.06-.13-.17a4.67 4.67 0 0 1 1.45-1.06c.08-.03.14.02.18.1Z"
-      style={{
-        fill: "#f8e5c5",
-      }}
-    />
-    <path
-      d="M109.43 202.24a.62.62 0 0 1 .57-.24.62.62 0 0 0 .7-.39.54.54 0 0 1-.61.76c-.22-.01-.44-.08-.66-.13Z"
-      style={{
-        fill: "#f0bb43",
-      }}
-    />
-    <path
-      d="M117.23 195.23a42 42 0 0 1-3.14 3.38.63.63 0 0 1 .23-.54l2.41-2.61c.14-.15.27-.29.5-.23Z"
-      style={{
-        fill: "#ac6403",
-      }}
-    />
-    <path
-      className="in"
-      d="M117.94 194.85h.16c.13 0 .35-.07.34.18s-.2.2-.35.2h-.27a.53.53 0 0 1 .12-.38Z"
-    />
-    <path
-      d="M117.83 194.32c.28-.29.47-.44.83-.27a.93.93 0 0 1-.83.27Z"
-      style={{
-        fill: "#a16202",
-      }}
-    />
-    <path
-      className="in"
-      d="M119.1 194.24c0-.16 0-.32-.12-.43.53-.94.88-2 1.35-3a1.76 1.76 0 0 0 .08-.82c.2.05.08.3.17.38.36 0 0-.69.44-.64a2.35 2.35 0 0 1-.18 1.17c-.27.67-.54 1.34-.8 2-.31.09-.28.43-.41.65s-.15.69-.53.69Z"
-    />
-    <path
-      d="M119.1 194.24c.43-.35.39-1 .76-1.41 0 0 0-.28.16-.06v.18a6.36 6.36 0 0 1-.62 1.41l-.33.11Z"
-      style={{
-        fill: "#f6e5c1",
-      }}
-    />
-    <path className="dg" d="m118 213.63-.51.86-.1-.24c.17-.25.26-.63.61-.62Z" />
-    <path
-      className="iu"
-      d="M104.94 204.08c0-.19.22-.38.06-.53s-.35 0-.52.11-.15.14-.23 0a1.52 1.52 0 0 1 1-.49c-.12.39.13.92-.37 1.15.05-.07.05-.16.06-.24ZM100.34 204.93h.48c.59.27 1.21.06 1.82.17a27.66 27.66 0 0 1-3.19.08c.34-.29.66-.03.89-.25Z"
-    />
-    <path
-      d="M119.34 205.18v1l-.24.47a6.55 6.55 0 0 1 .17-1.88c.24.05.07.23.07.41Z"
-      style={{
-        fill: "#fbf0d0",
-      }}
-    />
-    <path
-      className="ej"
-      d="M119.34 205.18c0-.14 0-.3-.12-.39.1-.72.13-.79.49-1-.21.43-.11.97-.37 1.39Z"
-    />
-    <path
-      className="m"
-      d="M176.21 166.74s-6.48-2.16-11.87 1.62c0 0-8.1 12.42-9.72 11.88l-5.7 3s11.64-1.88 13.8-4.58 3.77-6.48 8.63-9.18ZM182.15 163s3.78-2.7 9.18 0 9.72 4.32 9.72 4.32-14.58-4.32-16.74-2.7-2.16-1.62-2.16-1.62Z"
-    />
-  </svg>
-);
-
-export default SvgPotato;
